% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASLdata}
\alias{ASLdata}
\title{An example dataset formatted as subject level VAD}
\format{
A data frame with 80 obs. of  4 variables:
\describe{
\item{SUBJID}{}
\item{SEX}{}
\item{AGE}{}
\item{ARM}{}
}
}
\usage{
ASLdata
}
\description{
An example dataset formatted as subject level VAD
}
\keyword{datasets}
