% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATEdata}
\alias{ATEdata}
\title{An example dataset formatted as time to event efficacy VAD}
\format{
A data frame with 80 obs. of  4 variables:
}
\usage{
ATEdata
}
\description{
An example dataset formatted as time to event efficacy VAD
}
\keyword{datasets}
