% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsPredprob.R
\name{boundsPredprob}
\alias{boundsPredprob}
\title{Decision cutpoints for boundary (based on predictive probability) for Decision 1 rule.}
\usage{
boundsPredprob(looks, Nmax = max(looks), p0, tT, phiL, phiU, parE)
}
\arguments{
\item{Nmax}{(\code{number}):\cr maximum number of patients at the end of the trial in the \code{E} group.}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{parE}{(\code{numeric}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}
}
\value{
A matrix for each same size in \code{looks}. For each sample size, the following is returned:
\itemize{
\item \code{xL} : the maximum number of responses that meet the futility.
threshold
\item \code{pL} : response rate corresponding to \code{xL}.
\item \code{ppL} : predictive probability corresponding to \code{xL}
\item \code{postL}: posterior probability corresponding to \code{xL}.
\item \code{Ucil} : upper bound of one sided 95\% CI for the response rate based on an
exact binomial test.
\item \code{xU} : the minimal number of responses that meet the efficacy threshold.
\item \code{pU} : response rate corresponding to \code{xU}.
\item \code{postL}: posterior probability corresponding to \code{xU}.
\item \code{ppU} : predictive probability corresponding to \code{xU}
\item \code{LciU} : lower bound of one sided 95\% CI for the response rate based on exact
binomial test.
}
}
\description{
This function is used to identify the efficacy boundary and futility
boundary based on the following rules:
Efficacy boundary: find minimum x (xU) where
Pr(Pr(P > p0 | x, Y, a, b) >= tT | x) >= phiU,
Futility boundary: find maximum x (xL) where
Pr(Pr(P > p0 | x, Y, a, b) >= tT | x) =< phiL
}
\examples{
# 40 pts trial with interim looks after each 10 patients.
# Final efficacy decision if more than 80\% probability to be above 20\% ORR,
# Final futility decision otherwise.
# Interim efficacy decision if more than 90\% predictive probability reach this or
# Efficacy look Pr(Pr(P > p0 | x, Y, a, b) >= tT | x) >= phiU,
# Interim futility decision if less than 10\% predictive probability or
# Futility look Pr(Pr(P > p0 | x, Y, a, b) >= tT | x) =< phiL
# Uniform prior (i.e. beta(1, 1)) on the ORR:
boundsPredprob(
  nvec = c(10, 20, 30, 40),
  p = 0.20,
  tT = 0.80,
  phiL = 0.10,
  phiU = 0.90,
  a = 1,
  b = 1
)
# From this we see e.g. that at the first IA at 10 patients, we would stop for futility
# if no patient responded, and for efficacy if 4 or more patients responded.
}
\keyword{graphics}
