% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{dbetabinomMix}
\alias{dbetabinomMix}
\title{Beta-mixture-binomial density function}
\usage{
dbetabinomMix(x, m, par, weights, log = FALSE)
}
\arguments{
\item{x}{number of successes}

\item{m}{number of trials}

\item{par}{the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components}

\item{weights}{the mixture weights of the beta mixture prior}

\item{log}{return the log value? (not default)}
}
\value{
The (log) density values of the mixture of beta-binomial distributions at \code{x}.
}
\description{
Calculates the density function for a mixture of beta-binomial distributions.
}
\details{
Note that \code{x} can be a vector.
}
