% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiff.R
\name{dbetadiff}
\alias{dbetadiff}
\title{The Difference between Two Beta Distributions}
\usage{
dbetadiff(
  z,
  parY,
  parX,
  eps = .Machine$double.eps,
  rel.tol = .Machine$double.eps^0.1
)
}
\arguments{
\item{z}{(\code{numeric}):\cr vector of differences between Control and Treatment arms such that \code{Z = Y-X}}

\item{parY}{(\code{numeric}):\cr two parameters of \code{Y}'s Beta distribution (Treatment)}

\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}

\item{eps}{(\code{number}):\cr lowest floating point number as lower bound of integration}

\item{rel.tol}{(\code{number}):\cr used in \code{\link[stats:integrate]{stats::integrate()}}}
}
\value{
The density values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The Probability Density Function of the difference of two Beta Distributions.

Density, distribution function and quantile function for
the distribution of the difference of two Beta distributions with parameters \code{parX} and \code{parY}.
We denote \code{X} and \code{Y} as two random variables representing the response rate of Control and Treatment
group respectively. The assignment of Control and Treatment is practically interchangeable.
We denote \code{Z} as the difference between two groups such that \code{Z = Y-X}.
}
\note{
\code{X} and \code{Y} can be either Control or Treatment and \code{Z = X-Y}, subject to assumptions.
}
\examples{
# The following examples use these parameters:
parX <- c(1, 52) # Control group parameters
parY <- c(5.5, 20.5) # Treatment group parameters

# The difference between Control and Treatment is denoted as z.
z <- seq(from = -1, to = 1, length = 100)
plot(z, dbetadiff(z, parY = parY, parX = parX),
  type = "l"
)

# Calculate probability of Go, if difference was at least 15\%.
test <- integrate(
  f = dbetadiff,
  parY = parY,
  parX = parX,
  lower = 0.15,
  upper = 1
)
str(test)
test$value

# Calculate probability of Stop, if difference was at most 50\%.
integrate(
  f = dbetadiff,
  parY = parY,
  parX = parX,
  lower = -1,
  upper = 0.5
)
# Explanation: given density of the difference between two beta distributions
# is P( difference < 0.5 | dbetadiff ), and
# if the rule is STOP when difference < 0.5, then
# P( STOP | dbetadiff ) is high, or approximately 99 \%
# therefore if the difference was at most 50\%, then
# we will stop most of the time.
}
