% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{get_decision}
\alias{get_decision}
\title{get_decision}
\usage{
get_decision(nnr, response, truep, p0, p1, parE = c(1, 1), nnE, nnF, tL, tU)
}
\arguments{
\item{response}{(\code{numeric}):\cr A numeric of bernoulli successes based on \code{size_look},}

\item{truep}{(\code{number}):\cr assumed true response rate nor true rate (scenario).}

\item{p0}{(\code{number}):\cr lower efficacy threshold of response rate.}

\item{p1}{(\code{number}):\cr upper efficacy threshold of response rate.}

\item{parE}{(\code{numeric}):\cr beta parameters for the prior on the treatment proportion.}

\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for futility decision if different from Efficacy decision.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{<parse error>}{(\if{html}{\out{<parse error>}}):\cr \if{html}{\out{<parse error>}}}
}
\value{
A list of the following objects :
\itemize{
\item \code{decision} : resulting numeric of decision, one of \code{TRUE} for GO, \code{FALSE}for STOP, \code{NA} for Gray zone
\item \code{all_sizes} : resulting numeric of look size, anything below maximum
look size is an indicated interim, futility or efficacy or both
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A helper function for \code{ocPostprob} to generate numeric of decisions \code{decisions} and random looks \code{all_sizes}.
}
