% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{get_distance}
\alias{get_distance}
\title{Operating Characteristics for Posterior Probability method}
\usage{
get_distance(nn)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for efficacy decision. If different for futility decision,
specify in \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate nor true rate (scenario).}

\item{p0}{(\code{number}):\cr lower efficacy threshold of response rate.}

\item{p1}{(\code{number}):\cr upper efficacy threshold of response rate.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{parE}{(\code{numeric}):\cr beta parameters for the prior on the treatment proportion.}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{wiggle}{(\code{logical}):\cr generate random look locations (not default)
if \code{TRUE}, specify \code{dist} (see @details)}

\item{dist}{(\code{numeric} or \code{NULL}#TODO ( was dl)... check Roxytypes):\cr distance for random looks around the look locations in \code{nn}.
If \code{NULL}, only one location look will be set at nnE or nnF or n}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for futility decision. If different for futility decision,
specify in \code{nnF}.
\subsection{From helper functions}{
}}

\item{nnrE}{(\code{numeric}):\cr same as \code{nnE} but if wiggle room and distance applied.}

\item{nnrF}{(\code{numeric}):\cr same as \code{nnF} but if wiggle room and distance applied.}
}
\value{
A list with the following elements:
\itemize{
\item \code{oc}: matrix with operating characteristics (see Details section)
Decision: vector of the decisions made in the simulated trials
(\code{TRUE} for success, \code{FALSE} for failure, \code{NA} for no
decision)
SampleSize: vector of the sample sizes in the simulated trials
\item \code{nn}: vector of look locations that was supplied
\item \code{nnE}: vector of efficacy look locations
\item \code{nnF}: vector of futility look locations
\item \code{params}: multiple parameters
\item \code{Decision} : resulting decision, one of \code{TRUE} for GO, \code{FALSE}for STOP, \code{NA} for Gray zone
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate operating characteristics for posterior probability method.

It is assumed that the true response rate is \code{truep}.
The trial is stopped for efficacy if the posterior probability to be
above \code{p1} is larger than \code{tU}, and stopped for futility if the posterior
probability to be below \code{p0} is larger than \code{tL}:

Stop criteria for Efficacy : \code{P_E(p > p1) > tU}

Stop criteria for Futility : \code{P_E(p < p0) > tL}

Resulting Operating Characteristics include the following:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability to decide for Futility early
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: Probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\details{
\subsection{About arguments}{

\code{ExpectedN} is an average of the simulated sample sizes.
If \code{wiggle = TRUE}, one can specify \code{dist}, though the algorithm will generate it if \code{dist = NULL}
If \code{nnF = NULL}, no Futility or decision to Stop will be analysed. Note that \code{nnF = c(0)} is equivalent.
As default, \code{nnF} is set to the identical looks of \code{nnE}, and if \code{wiggle = TRUE}, all looks are the same, e.g.
\code{nnE = nnF} when wiggle and distance is applied.
}

\subsection{About helper function}{

\code{get_distance} inputs \code{dist} into \code{get_looks} and thereafter contributes to arguments in \code{get_decision}.
Finally, \code{get_oc} generates a list of parameters such as \code{decisions}, \code{all_sizes} and operating characteristics (oc).
}
}
\examples{
# operating characteristics for posterior probability method

# design details (example)
# multiple looks @ 10, 20, 30 patients
# True response rate of the treatment group=40\%
# stop for futility: P(response rate < 20\% )> 60\%
# s top for efficacy: P(response rate > 30\% )> 80\%
# prior of treatment arm parE= Beta(1,1)
res1 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000
)
res1$oc

# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:
res2 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE
)
res2$oc

# now d is specified:
res3 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE, d = 5
)
res3$oc

# finally, we can also have separate specification of efficacy and
# futility analyses. E.g. here have futility decisions only at the final analysis:
res4 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE, d = 5,
  nnF = c(30)
)
res4$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.
}
