% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{get_looks}
\alias{get_looks}
\title{get_looks}
\usage{
get_looks(dist, nnE, nnF)
}
\arguments{
\item{dist}{(\code{numeric}):\cr Distance generated from \code{get_distance} in a numeric of at least one element.}

\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for futility decision if different from Efficacy decision.}
}
\value{
A numeric of looks with outputs from \code{get_distance} randomly added to looks.
}
\description{
A helper function for \code{ocPostprob} that applies the numeric element of distance to looks \code{nn}.
}
\examples{
examples / ocPostProb
}
