% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{get_oc}
\alias{get_oc}
\title{Title}
\usage{
get_oc(all_sizes, nnr, decision, nnrE, nnrF)
}
\arguments{
\item{nnrF}{}

\item{<parse error>}{(\if{html}{\out{<parse error>}}):\cr \if{html}{\out{<parse error>}}}
}
\description{
Title
}
