% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{getBetamixPost}
\alias{getBetamixPost}
\title{Computes the posterior parameters of a beta mixture}
\usage{
getBetamixPost(x, n, par, weights)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of patients}

\item{par}{the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components}

\item{weights}{the mixture weights}
}
\value{
a list with the updated beta parameters and weights
}
\description{
Computes the posterior parameters of a beta mixture
}
\examples{
## example from Lee and Liu:
getBetamixPost(x=16, n=23, par=t(c(0.6, 0.4)), weights=1)

getBetamixPost(x=16, n=23,
               par=
               rbind(c(0.6, 0.4),
                     c(1, 1)),
               weights=c(0.6, 0.4))

}
