% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprobDist.R
\name{h_decision_one_predprobDist}
\alias{h_decision_one_predprobDist}
\title{Generating random decision and sample size looks for \code{decision1 == TRUE} or default option}
\usage{
h_decision_one_predprobDist(
  nnE,
  nnF,
  nnr,
  truep,
  xS,
  nS,
  parE = c(1, 1),
  parS = c(1, 1),
  tT,
  phiU,
  phiL,
  deltaE,
  deltaF,
  weights,
  weightsS,
  relativeDelta = FALSE
)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If \code{0} or \code{NULL} and
\code{length(nnE) = 1} then no Efficacy looks are performed.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If \code{0} or \code{NULL} and
\code{length(nnF) = 1} then no Futility looks are performed.}

\item{nnr}{(\code{numeric}):\cr union of \code{nnE} and \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{xS}{(\code{number}):\cr number of successes in the \code{S} group at interim.}

\item{nS}{(\code{number}):\cr number of patients in the \code{E} group at interim.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution for the \code{S} group.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}

\item{deltaE}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Efficacy looks only.
Note that this can also be negative as well.}

\item{deltaF}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Futility looks only.
Note that this can also be negative as well.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior.}

\item{weightsS}{(\code{numeric}):\cr weights for the \code{S} group.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients. See \code{\link[=postprobDist]{postprobDist()}}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{decision} : decision \code{flag} with \code{TRUE} for Go, \code{FALSE} for Stop, \code{NA} for Gray zone.
\item \code{all_sizes} : resulting number of look size, anything below maximum
look size is an indicated interim, Futility or Efficacy or both.
}
}
\description{
A helper function for \code{\link[=ocPredprobDist]{ocPredprobDist()}} to generate numeric of decisions \code{decisions} and
random looks \code{all_sizes} for \code{decision1 == TRUE}.
}
\keyword{internal}
