% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocRctPredprobDist.R
\name{h_decision_one_RctpredprobDist}
\alias{h_decision_one_RctpredprobDist}
\title{Generating random decision and sample size looks}
\usage{
h_decision_one_RctpredprobDist(
  nnr,
  nnE,
  nnF,
  pE,
  pS,
  parE,
  parS,
  tT,
  phiU,
  phiL,
  deltaE,
  deltaF,
  weights,
  weightsS,
  relativeDelta = FALSE,
  randRatio = 1,
  orig_nnr
)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE}and \code{nnF}.}

\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{S} group. See details.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}

\item{deltaE}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Efficacy looks only.
Note that this can also be negative as well.}

\item{deltaF}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Futility looks only.
Note that this can also be negative as well.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior.}

\item{weightsS}{(\code{numeric}):\cr weights for the \code{S} group.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients
is included as the margin between treatment and control group. See note.}

\item{orig_nnr}{(\code{numeric}):\cr unique values of original, user-input values of \code{nnE} and \code{nnF}, defined by user-facing
function and aid to \code{plotOc()}'s x axis input when \code{wiggle = TRUE}.}
}
\value{
A list with the following elements :
\itemize{
\item \code{decision} : decision \code{flag} with \code{TRUE} for Go, \code{FALSE} for Stop, \code{NA} for Gray zone.
\item \code{all_sizes} : resulting sample size.
\item \code{nActive} : number of patients in the active arm.
\item \code{nControl} : number of patients in the control arm.
}
}
\description{
A helper function for \code{\link[=ocRctPredprobDist]{ocRctPredprobDist()}} to generate a single trial decision and sample size.
}
\keyword{internal}
