% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{h_dist0}
\alias{h_dist0}
\title{Generating random distance for \code{h_get_distance} helper function.}
\usage{
h_dist0(nn)
}
\arguments{
\item{nn}{(\verb{number or numeric}):\cr the union of \code{nnE} and \code{nnF} if futility analysis or looks are supplied.}
}
\value{
A number providing the wiggle distance.
}
\description{
A helper function for \code{h_get_distance} by first calculating non-overlapping
distance between looks if looks are of minimum length of 2 elements.
}
\keyword{internal}
