% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{h_get_bounds}
\alias{h_get_bounds}
\title{Generating bounds for the Integration of Beta Mixture Posterior}
\usage{
h_get_bounds(controlBetamixPost)
}
\arguments{
\item{controlBetamixPost}{(\code{list}):\cr a list of posterior parameters of a beta-mixture-binomial distribution with generic names
\code{par} and \code{weights}. See \verb{[getBetaMix()]}.}
}
\value{
Integrand function
}
\description{
Using the quantile of the Beta Mixture Distribution from parameters given by standard of care \code{S} or
experimental group \code{E} to determine bounds as inputs to \verb{[stats::integrate()]}.
}
\keyword{internal}
