% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOc.R
\name{h_get_dataframe_oc}
\alias{h_get_dataframe_oc}
\title{Helper function for simulation result as input for \code{plotOc()}}
\usage{
h_get_dataframe_oc(decision, sample_size, all_looks)
}
\arguments{
\item{decision}{(\code{numeric}):\cr Go, Stop or Gray Zone decisions of all looks simulated \code{length(sim)} times.}

\item{all_looks}{(\code{numeric}):\cr original looks before adjustment by \code{wiggle = TRUE}, if applied.}
}
\description{
Data frame input for bar plot for simulated results of :
\itemize{
\item \verb{[ocPostprob()]}
\item \verb{[ocPostprobDist()]}
\item \verb{[ocPostpred()]}
\item \verb{[ocPostpredDist()]}
\item \verb{[ocRctPostprobDist()]}
\item \verb{[ocRctPredprobDist()]}
}
}
\keyword{internal}
