% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{h_get_decision_one_predprob}
\alias{h_get_decision_one_predprob}
\title{Generating random decision and sample size looks for \code{decision1 == TRUE} or default option}
\usage{
h_get_decision_one_predprob(
  nnr,
  truep,
  p0,
  parE = c(1, 1),
  nnE,
  nnF,
  tT,
  phiU,
  phiL
)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE}and \code{nnF} from \code{ocPredprob}.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}
}
\value{
A list with the following elements:
\itemize{
\item \code{decision} : decision \code{flag} with \code{TRUE} for Go, \code{FALSE} for Stop, \code{NA} for Gray zone.
\item \code{all_sizes} : resulting number of look size, anything below maximum
look size is an indicated interim, Futility or Efficacy or both.
}
}
\description{
A helper function for \code{ocPredprob} to generate numeric of decisions \code{decisions} and
random looks \code{all_sizes} for \code{decision1 == TRUE}.
}
\keyword{internal}
