% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{h_get_decision_two_predprob}
\alias{h_get_decision_two_predprob}
\title{Title}
\usage{
h_get_decision_two_predprob(
  nnr,
  truep,
  p0,
  p1,
  parE = c(1, 1),
  nnE,
  nnF,
  tT,
  tF,
  phiFu,
  phiU
)
}
\arguments{
\item{<parse error>}{(\if{html}{\out{<parse error>}}):\cr \if{html}{\out{<parse error>}}}
}
\description{
Title
}
