% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprobDist.R
\name{h_get_decisionDist}
\alias{h_get_decisionDist}
\title{Evaluate Posteriors based on Efficacy and Futility thresholds in Two-armed trials}
\usage{
h_get_decisionDist(
  nnr,
  nnrE,
  nnrF,
  truep,
  parE = c(1, 1),
  parS = c(1, 1),
  tL,
  tU,
  deltaE,
  deltaF,
  relativeDelta
)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE}and \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{S} group. See details.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{deltaE}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group.}

\item{deltaF}{(\code{number}):\cr delta for efficacy: \code{P(P_E > P_S + deltaE)} should be large
to stop for efficacy.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients. See note.}
}
\description{
A helper function adapted from \code{h_get_decision()} from \code{ocPostprob.R} by
replacing the internal use of \code{postprob()} with \code{postprobDist()}
to generate numeric of \code{decisions} and random looks \code{all_sizes}.
}
\keyword{internal}
