% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{h_get_distance}
\alias{h_get_distance}
\title{Generating random distance in given looks for sample sizes for Efficacy and Futility.}
\usage{
h_get_distance(nn)
}
\arguments{
\item{nn}{(\verb{number or numeric}):\cr the union of \code{nnE} and \code{nnF}, if futility analysis or looks are supplied.}
}
\value{
A numeric with \code{length(nn)-1} elements.
}
\description{
A helper function for \code{\link[=ocPostprob]{ocPostprob()}} to generate random distance's wiggle room around looks \code{nn}.
}
\keyword{internal}
