% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{h_get_oc}
\alias{h_get_oc}
\title{Creating list for operating characteristics.}
\usage{
h_get_oc(all_sizes, nnr, decision, nnrE, nnrF)
}
\arguments{
\item{all_sizes}{(\code{numeric}):\cr Sample sizes of all looks simulated \code{length(sim)} times if \code{dist} applied.}

\item{nnr}{(\code{numeric}):\cr union of \code{nnE}and \code{nnF}.}

\item{decision}{(\code{numeric}):\cr Go, Stop or Gray Zone decisions of all looks simulated \code{length(sim)} times.}

\item{nnrE}{(\code{numeric}):\cr Looks with random distance, if applied on \code{nnE}.}

\item{nnrF}{(\code{numeric}):\cr Looks with random distance, if applied on \code{nnF}.}
}
\value{
A list of results containing :
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability to decide for Futility early
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: Probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\description{
Generates operating characteristics.
}
\examples{
examples / ocPostprob.R
}
\keyword{internal}
