% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{h_get_oc_predprob}
\alias{h_get_oc_predprob}
\title{Creating list for operating characteristics of \code{ocPredprob}}
\usage{
h_get_oc_predprob(all_sizes, nnr, decision)
}
\arguments{
\item{all_sizes}{(\code{numeric}):\cr Sample sizes of all trials.}

\item{nnr}{(\code{numeric}):\cr union of \code{nnE} and \code{nnF}.}

\item{decision}{(\code{numeric}):\cr Go (\code{TRUE}), Stop (\code{FALSE}) or Gray Zone (\code{NA}) decisions of all trials.}
}
\value{
A list of results containing :
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability of Gray Zone decision
}
}
\description{
A helper function for \code{\link[=ocPredprob]{ocPredprob()}} to generate numeric of decisions \code{decisions} and
random looks \code{all_sizes} for \code{decision1 == TRUE}.
}
\keyword{internal}
