% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocRctPostprobDist.R
\name{h_get_oc_rct}
\alias{h_get_oc_rct}
\title{Creating list for operating characteristics}
\usage{
h_get_oc_rct(all_sizes, Nmax, nActive, nControl, decision)
}
\arguments{
\item{Nmax}{(\code{number}):\cr The max sample size or the sample size of final look.}
}
\value{
A list of results containing :
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials in both treatment and SOC group
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
\item \code{ExpectedNactive} : the mean of the number of patients in treatment arm
\item \code{ExpectedNcontrol}: the mean of the number of patients in control arm
}
}
\description{
Generates operating characteristics for \code{\link[=ocRctPostprobDist]{ocRctPostprobDist()}}.
}
\keyword{internal}
