% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{h_getBetamixPost}
\alias{h_getBetamixPost}
\title{Compute Beta-Mixture-Binomial Posterior Distribution}
\usage{
h_getBetamixPost(x, n, par, weights)
}
\arguments{
\item{x}{(``):\cr number of successes}

\item{n}{(``):\cr number of patients}

\item{par}{(``):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components}

\item{weights}{(``):\cr the mixture weights}
}
\value{
A list with the updated beta parameters and weights
}
\description{
Computes the posterior parameters of a beta-mixture-binomial distribution.
}
\keyword{internal}
