% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{h_integrand_relDelta}
\alias{h_integrand_relDelta}
\title{The Posterior Beta Mixture Integrand based on \code{delta}}
\usage{
h_integrand_relDelta(p_s, delta, x, betamixPost, controlBetamixPost)
}
\arguments{
\item{p}{(\code{number}):\cr probability of success or response rate of standard of care or \code{SOC} group.}
}
\value{
An R function that is an argument for \verb{[stats::integrate()]}.
}
\description{
The helper function to generate Integrand function when  \verb{relative Delta = TRUE}.
}
\keyword{internal}
