% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprobDist.R
\name{h_predprobdist}
\alias{h_predprobdist}
\title{The predictive probability of success in two-arm studies.}
\usage{
h_predprobdist(
  NmaxControl,
  Nmax,
  nS,
  xS,
  parS,
  parE,
  weights,
  weightsS,
  x,
  mE,
  density_y,
  delta,
  relativeDelta
)
}
\value{
A list with predictive probability of success
}
\description{
The helper function to generate the predictive probability of success
based on the difference in treatment group (\code{E}) and control or
standard of care (\code{SOC}) group.
}
\keyword{internal}
