% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprobDist.R
\name{h_predprobdist_single_arm}
\alias{h_predprobdist_single_arm}
\title{The predictive probability of success in single arm studies.}
\usage{
h_predprobdist_single_arm(
  x,
  Nmax,
  delta,
  relativeDelta,
  parE,
  weights,
  parS,
  weightsS,
  thetaT,
  density,
  mE
)
}
\arguments{
\item{x}{(\code{number}):\cr number of successes in the treatment group at interim.}

\item{Nmax}{(\code{number}):\cr maximum number of patients at final analysis.}

\item{delta}{(\code{number}):\cr difference between response rates to be met.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the SOC non-responding patients. See note.}

\item{parE}{(\code{numeric}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the treatment group. See details.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}

\item{parS}{(\code{numeric}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution for the control group.}

\item{weightsS}{(\code{numeric}):\cr weights for the SOC group.}

\item{thetaT}{(\code{number}):\cr threshold on the probability to be used.}

\item{density}{(\code{numeric}):\cr the beta binomial density for future success in \code{Nmax-n} patients in the treatment group.}

\item{mE}{(\code{number}):\cr number of successes in the remaining \code{Nmax-n} number of patients in the treatment group.}
}
\description{
The helper function to generate The predictive probability of success when \code{NmaxControl == 0}.
}
