% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logit}
\alias{logit}
\title{The logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{vector of probabilities}
}
\value{
logit(x) calculates logit(x)=log(x/(1-x))
}
\description{
The logit function
}
\examples{
## If x = 0.2
## then logit(x) = log(0.2/(1-0.2)) = -1.386294
##
x <- 0.2
logit(x)

## Can also be a vector of probabilities
##
x <- seq(0.1, 0.9, 0.1)
logit(x)
}
\keyword{programming}
