% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{myPlot}
\alias{myPlot}
\title{Plot the Beta distribution}
\usage{
myPlot(alpha, beta, ...)
}
\arguments{
\item{alpha}{first parameter of the Beta distribution}

\item{beta}{second parameter of the Beta distribution}

\item{\dots}{additional arguments to \code{plot}}
}
\value{
nothing, only produces the plot as side effect
}
\description{
This function will plot the PDF of a beta distribution
}
\examples{
myPlot(1, 2)
myPlot(1, 1)
}
\keyword{graphics}
