% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprobDist.R
\name{ocPostprobDist}
\alias{ocPostprobDist}
\title{Calculate operating characteristics for posterior probability method
with beta prior on SOC}
\usage{
ocPostprobDist(
  nn,
  p,
  deltaE,
  deltaF,
  relativeDelta = FALSE,
  tL,
  tU,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)
}
\arguments{
\item{nn}{vector of look locations for efficacy
(if futility looks should be different, please specify also \code{nnF})}

\item{p}{true rate (scenario)}

\item{deltaE}{delta for efficacy: P(pE > pS + deltaE) should be large
to stop for efficacy}

\item{deltaF}{delta for futility: P(pE < pS - deltaF) should be large to
stop for futility}

\item{relativeDelta}{see \code{\link{postprobDist}}}

\item{tL}{probability threshold for being below control - deltaF}

\item{tU}{probability threshold for being above control + deltaE}

\item{parE}{beta parameters for the prior on the treatment proportion}

\item{parS}{beta parameters for the prior on the control proportion}

\item{ns}{number of simulations}

\item{nr}{generate random look locations? (not default)}

\item{d}{distance for random looks around the look locations in \code{nn}}

\item{nnF}{vector of look locations for futility
(default: same as efficacy)}
}
\value{
A list with the following elements:
oc: matrix with operating characteristics (see Details section)
Decision: vector of the decisions made in the simulated trials
(\code{TRUE} for success, \code{FALSE} for failure, \code{NA} for no
decision)
SampleSize: vector of the sample sizes in the simulated trials
nn: vector of look locations
nnE: vector of efficacy look locations
nnF: vector of futility look locations
todo: would we like to return nnr instead, the actual look locations?
params: input parameters for this function
}
\description{
The trial is stopped for efficacy if the posterior probability to be at
least deltaE better than the control is larger than tU, and stopped for
futility if the posterior probability to be at least deltaF worse than the
control is larger than tL. Otherwise the trial is continued, and at the
maximum sample size it may happen that no decision is made ("gray zone").
}
\details{
Returned operating characteristics in a matrix include:
ExpectedN: expected number of patients in the trials
PrStopEarly: probability to stop the trial early (before reaching the
maximum sample size)
PrEarlyEff: probability to decide for efficacy early
PrEarlyFut: probability to decide for futility early
PrEfficacy: probability to decide for efficacy
PrFutility: probability to decide for futility
PrGrayZone: probability of no decision at the end ("gray zone")
}
\examples{
# operating characteristics for posterior probability method with beta prior on SOC

# design details
# multiple looks @ 10, 20, 30 patietns
# True response rate of the treatment group=0.4

# stop for efficacy (deltaE): P(pE > pS + deltaE) >tU
# stop for futility (deltaF): P(pE < pS - deltaF) >tL
# where pE is the response rate of treatment group, pS is the response rate of
# control group. Both of them are random variables.


res1 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100
)

res1$oc


# generate random look locations around 10,20,30 patients
# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:
res2 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100, nr = TRUE
)

res2$oc

# specify the distance for random looks around the look locations in nn (d=5 for illustration)
res3 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100,
  nr = TRUE,
  d = 5
)

res3$oc


# finally, we can also have separate specification of efficacy and futility analyses. E.g.
# here have futility decisions only at the final analysis:
res4 <- ocPostprobDist(
  nn = c(10, 20, 30), p = 0.4, deltaE = 0.1, deltaF = -0.1, tL = 0.6, tU = 0.6,
  parE = c(1, 1), parS = c(5, 25), ns = 100, nr = TRUE, d = 5, nnF = c(30)
)
res4$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.
}
