% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{ocPredprob}
\alias{ocPredprob}
\title{Operating Characteristics for Predictive Probability method}
\usage{
ocPredprob(
  nnE,
  truep,
  p0,
  phiU,
  p1 = p0,
  tT = 1 - tF,
  tF = 1 - tT,
  phiL = 1 - phiFu,
  phiFu = 1 - phiL,
  parE = c(1, 1),
  sim = 50000,
  wiggle = FALSE,
  nnF = nnE,
  decision1 = TRUE
)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If \code{0} or \code{NULL} and
\code{length(nnE) = 1} then no Efficacy looks are performed.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{p1}{(\code{number}):\cr upper Futility threshold of response rate.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{tF}{(\code{number}):\cr threshold of which assumed \code{truep} does not exceed threshold of \code{p1}.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}

\item{phiFu}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{wiggle}{(\code{flag}):\cr generate random look locations (not default). If \code{wiggle = TRUE} and \code{nnE = nnF}, then all wiggled
looks are the same between \code{nnE} and \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If \code{0} or \code{NULL} and
\code{length(nnF) = 1} then no Futility looks are performed.}

\item{decision1}{(\code{flag}):\cr Flag if \code{decision1 = TRUE} then Decision 1 rules will be used, otherwise Decision 2 rules will be used.}
}
\value{
A list with the following elements:
\itemize{
\item \code{oc}: matrix with operating characteristics with the following details:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
\item \code{Decision} : numeric of results with \code{TRUE} as Go decision, \code{FALSE} as Stop and \code{NA} as gray zone.
\item \code{SampleSize} : numeric of sample sizes from \code{nnE} or \code{nnF} or both.
\item \code{wiggled_nnE} : user input for \code{nnE} with random distance applied.
\item \code{wiggled_nnF} : user input for \code{nnF} with random distance applied.
\item \code{wiggled_dist} : magnitude of random distance applied in order of input looks.
\item \code{params} : all user input arguments.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate operating characteristics for predictive probability method.

It is assumed that the true response rate is \code{truep}. Trials can stop for Futility
or Efficacy. Trials can also stop at Interim or Final for Futility or Efficacy.

There are two variations of decision rule, Decision 1 and Decision 2, to
evaluate decision at Interim or Final, for Futility or Efficacy. Decision 1 is
used when \code{decision1 == TRUE} which is the default setting.
\subsection{Decision 1:}{

The criteria for Decision 1 for Interim looks are :
\itemize{
\item interim GO =  P(successful trial at final) > phiU
\item interim STOP = P(successful trial at final) < phiL
}

The criteria for Decision 1 for Final looks are:
\itemize{
\item Final GO = P( response rate > p0 | data) > tT
\item Final STOP = P( response rate > p0 | data ) < tT
}
}

\subsection{Decision 2:}{

The criteria for Decision 2 for Interim looks are :
\itemize{
\item Interim GO : P ( success at final) > phiU
\item Interim STOP : P (failure at final ) > phiFu
}

The criteria for Decision 2 for Futility looks are :
\itemize{
\item Final GO = P( response rate > p0) > tT
\item Final STOP = P( response rate  < p1) > tF
}
}
}
\examples{
# Here we illustrate an example for Decision 1 with the following assumptions :
# True response rate or truep of the treatment group = 40\%
# The following are the Final Stop rules respectively :
# - Final look for Efficacy: Pr( response rate > 25\% ) > 60\% or P(response rate > p0) > tT
# - Final look for Futility: Pr( response rate < 25\% ) < 60\% or P(response rate > p0) < tT
# - Interim look for Efficacy: Pr( success at final ) > 80\% or P(success at final) > phiU
# - Interim look for Futility: Pr( failure at final ) < 20\% or P(success at final) < phiL
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.

# Decision 1 with no wiggle.
set.seed(20)
result <- ocPredprob(
  nnE = c(10, 20),
  truep = 0.4,
  p0 = 0.25,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 100,
  wiggle = FALSE,
  decision1 = TRUE
)
result$oc

# Deision 1 with wiggle.
result <- ocPredprob(
  nnE = c(10, 20),
  truep = 0.4,
  p0 = 0.25,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = TRUE,
  nnF = c(10, 20),
  decision1 = TRUE
)
result$oc

# Decision 1 with separate Futility and Efficacy looks at interim and final without wiggle.
result <- ocPredprob(
  nnE = c(10, 25, 30),
  truep = 0.4,
  p0 = 0.25,
  p1 = 0.2,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = FALSE,
  nnF = c(10, 15, 20),
  decision1 = TRUE
)
result$oc

# Decision 1 with separate Futility and Efficacy looks at interim and final with wiggle.
result <- ocPredprob(
  nnE = c(10, 25, 30),
  truep = 0.4,
  p0 = 0.25,
  p1 = 0.2,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = TRUE,
  nnF = c(10, 15, 20),
  decision1 = TRUE
)
result$oc

# Here we illustrate an example for Decision 2 with the following assumptions :
# True response rate or truep of the treatment group = 60\%
# The following are the Final Stop rules respectively :
# - Final look for Efficacy: Pr( response rate > 25\% ) > 60\% or P(response rate > p0) > tT
# - Final look for Futility: Pr( response rate < 25\% ) < 60\% or P(response rate < p1) > tF
# - Interim look for Efficacy: Pr( success at final ) > 80\% or P(success at final) > phiU
# - Interim look for Futility: Pr( failure at final ) > 80\% or P(failure at final) > phiFu
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.

# Decision 2 without wiggle.
result <- ocPredprob(
  nnE = c(10, 20),
  truep = 0.6,
  p0 = 0.25,
  p1 = 0.25,
  tT = 0.6,
  tF = 0.6,
  phiU = 0.8,
  phiFu = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = FALSE,
  nnF = c(10, 20),
  decision1 = FALSE
)
result$oc

# Decision 2 with wiggle and with Futility only at final with non-uniform beta prior parE.
result <- ocPredprob(
  nnE = c(10, 25, 30),
  truep = 0.6,
  p0 = 0.25,
  p1 = 0.25,
  tT = 0.6,
  tF = 0.6,
  phiL = 0.8,
  phiU = 0.8,
  parE = c(11, 19),
  sim = 50,
  wiggle = TRUE,
  nnF = 30,
  decision1 = FALSE
)
result$oc
}
