% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocRctPostprobDist.R
\name{ocRctPostprobDist}
\alias{ocRctPostprobDist}
\title{Calculate operating characteristics for RCT against SOC,
using the posterior probability method with beta priors}
\usage{
ocRctPostprobDist(
  nn,
  pE,
  pS,
  deltaE,
  deltaF,
  relativeDelta = FALSE,
  tL = 0.8,
  tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)
}
\arguments{
\item{nn}{vector of look locations for efficacy. Note that the maximum sample size is
derived as the maximum value (latest look) in this vector.
(if futility looks should be different, please specify also \code{nnF})}

\item{pE}{true rate in the treatment population}

\item{pS}{true rate in the standard of care population}

\item{deltaE}{delta for efficacy: P(pE > pS + deltaE) should be large
to stop for efficacy. Must be non-negative value between 0 and 1.}

\item{deltaF}{delta for futility: P(pE < pS - deltaF) should be large to
stop for futility. Must be numeric value between -1 and 1.
Can also be \code{NULL}. (see details)}

\item{relativeDelta}{see \code{\link{postprobDist}}}

\item{tL}{probability threshold for being below control - deltaF (default: 0.8)}

\item{tU}{probability threshold for being above control + deltaE (default: 0.8)}

\item{parE}{beta parameters for the prior on the control proportion}

\item{parS}{beta parameters for the prior on the treatment proportion}

\item{randRatio}{the randomization ratio (active vs. control) to be used.
default: 1.}

\item{ns}{number of simulations}

\item{nr}{generate random look locations? (not default)}

\item{d}{distance for random looks around the look locations in \code{nn}}

\item{nnF}{vector of look locations for futility
(default: same as efficacy)}
}
\value{
A list with the following elements:
oc: matrix with operating characteristics (see Details section)
Decision: vector of the decisions made in the simulated trials
(\code{TRUE} for success, \code{FALSE} for failure, \code{NA} for no
decision)
SampleSize: vector of the sample sizes in the simulated trials
SampleSizeActive: vector of the patients with treatment in the simulated trials
SampleSizeControl: vector of the patients with control in the simulated trials
nn: vector of look locations
nnE: vector of efficacy look locations
nnF: vector of futility look locations
todo: would we like to return nnr instead, the actual look locations?
params: input parameters for this function
}
\description{
The randomization works as follows. According to the randomization ratio and
the maximum sample size, patients are allocated to the treatment and SOC
arms. The number of patients in the active treatment arm is rounded to the
next higher integer. That is, the sequence of patients is determined from
the start, such that the number of patients in both arms is constant across
trial simulations.
}
\details{
The randomized controlled trial (RCT) is stopped for efficacy if the
posterior probability to be at least deltaE better than the control is
larger than tU, and stopped for futility if the posterior probability to be
at least deltaF worse than the control is larger than tL. Otherwise the
trial is continued, and at the maximum sample size it may happen that no
decision is made ("gray zone").

A variation can be requested when \code{deltaF} is set to \code{NULL}. Then
the futility decision is made when the probability to be deltaE better than
the control is lower than 1 - tL.

Returned operating characteristics in a matrix include:
ExpectedN: expected total number of patients in the trials
ExpectedNactive: expected number of patients with treatment
ExpectedNcontrol: expected number of patients with SOC
PrStopEarly: probability to stop the trial early (before reaching the
maximum sample size)
PrEfficacy: probability to decide for efficacy
PrFutility: probability to decide for futility
PrGrayZone: probability of no decision at the end ("gray zone")
}
\examples{
# note, deltaF has to be a value between 0 and 1 including the bounds, or NULL.

res <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, deltaE = 0.15, deltaF = 0.05,
  relativeDelta = FALSE,
  tL = 0.8, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL
)

res$oc


# separate specification of efficacy and futility analyses can be performed. E.g.
# here have futility decisions only at the final analysis:

res2 <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, deltaE = 0.15, deltaF = 0.05,
  relativeDelta = FALSE,
  tL = 0.8, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL, nnF = c(30)
)
res2$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.

# Also, when deltaF is set to NULL, the futility decision is made when the probability to
# be deltaE better than the control is lower than 1 - tL. i.e. Pr(pE-pS>deltaE)<1-tL
res3 <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, deltaE = 0.15, deltaF = NULL,
  relativeDelta = FALSE,
  tL = 0.2, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL
)
res3$oc

# As when deltaF is set to null, it sometimes makes sense for a futility decision
# to be made even when the treatment is better than the control, e.g. when that difference
# is small. This can be accomplished by specifying a negative value for deltaF.
res4 <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3,
  deltaE = 0.15, deltaF = -0.05,
  relativeDelta = FALSE,
  tL = 0.2, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL
)
res4$oc
}
