% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiff.R
\name{pbetadiff}
\alias{pbetadiff}
\title{Beta difference Cumulative Probability Function}
\usage{
pbetadiff(q, parY, parX)
}
\arguments{
\item{q}{(\code{number}):\cr vector of quantiles}

\item{parY}{(\code{numeric}):\cr two parameters of \code{Y}'s Beta distribution (Treatment)}

\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}
}
\value{
\code{pbetadiff} the distribution function
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the Cumulative Probability Function of the Beta difference for a given probability.
}
\examples{
# pbetadiff ----
parX <- c(1, 52)
parY <- c(5.5, 20.5)

# Calculate probability based on quantile `Q(Z) =< 0.122838`
pbetadiff(
  q = 0.122838,
  parY = parY,
  parX = parX
)

# Calculate probability based on quantile `Q(Z) =< 0.5`
pbetadiff(
  q = 0.5,
  parY = parY,
  parX = parX
)
}
