% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase1b-package.R
\docType{package}
\name{Phase1b-package}
\alias{Phase1b-package}
\alias{Phase1b}
\alias{phase1b}
\title{Calculations for decisions on Phase 1b clinical trials}
\description{
Phase 1b tools
}
\keyword{package}
