% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBeta}
\alias{plotBeta}
\title{Plot the Beta distribution}
\usage{
plotBeta(alpha, beta, ...)
}
\arguments{
\item{alpha}{(\code{number}):\cr first parameter of the Beta distribution}

\item{beta}{(\code{number}):\cr second parameter of the Beta distribution}
}
\value{
A beta distribution density plot
}
\description{
This function will plot the PDF of a beta distribution
}
\examples{
# plotBeta
plotBeta(alpha = 4, beta = 5)
plotBeta(alpha = 1, beta = 1)

# plotBetDiff
parX <- c(1, 52) # parameters of experimental arm
parY <- c(5.5, 20.5) # parameters of control or SOC
plotBetaDiff(
  parY = parY,
  parX = parX,
  Go_cut = 0.3,
  Stop_cut = 0.1,
  shade = TRUE,
  note = TRUE
)
}
\keyword{graphics}
