% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBetaDiff}
\alias{plotBetaDiff}
\title{Plot difference Between two Beta distributions}
\usage{
plotBetaDiff(
  parY,
  parX,
  Go_cut = 0.2,
  Stop_cut = 0.1,
  shade = TRUE,
  note = TRUE
)
}
\arguments{
\item{parY}{(\code{numeric}):\cr non-negative parameters of the treatment Beta distribution.}

\item{parX}{(\code{numeric}):\cr non-negative parameters of the historical control Beta distribution}

\item{shade}{(\code{flag}):\cr paint the two areas under the curve, default value = TRUE}

\item{note}{(\code{flag}):\cr show values of the colored area, default value = TRUE}

\item{cut_B}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{cut_W}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{...}{(``):\cr additional arguments to \code{ggplot()}}
}
\value{
a ggplot object
}
\description{
This function will plot the PDF of a difference between two Beta distributions
}
\examples{
# The beta distribution and acceptable bounds for
# a meaningful improvement of 0.20 and worsening of 0.05
parX <- c(1, 52) # parameters of experimental arm
parY <- c(5.5, 20.5) # parameters of control or SOC
plotBetaDiff(
  parY = parY,
  parX = parX,
  Go_cut = 0.3,
  Stop_cut = 0.1,
  shade = TRUE,
  note = TRUE
)
}
\keyword{graphics}
