% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBetaDist}
\alias{plotBetaDist}
\title{Plot Diff Between two Beta distributions}
\usage{
plotBetaDist(parY, parX, cut_B = 0.2, cut_W = 0.1, shade = 1, note = 1, ...)
}
\arguments{
\item{parY}{(\code{numeric}):\cr non-negative parameters of the treatment Beta distribution.}

\item{parX}{(\code{numeric}):\cr non-negative parameters of the historical control Beta distribution}

\item{cut_B}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{cut_W}{(\code{number}):\cr a poor improvement throshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{shade}{(``):\cr paint the two areas under the curve, default value=1 as "yes". other numbers stands for "no";}

\item{note}{(\code{number}):\cr show values of the colored area, default value=1 as "yes". other numbers stands for "no"}

\item{<parse error>}{(\if{html}{\out{<parse error>}}):\cr \if{html}{\out{<parse error>}}}
}
\value{
a ggplot object
}
\description{
This function will plot the PDF of a diffience between two Beta distributions
}
\examples{
myPlotDiff(c(5, 10), c(2, 5), 0.2, 0.05, 1, 0)
}
\keyword{graphics}
