% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBounds.R
\name{plotBounds}
\alias{plotBounds}
\title{Plot the boundary decision cutpoints}
\usage{
plotBounds(
  z,
  area = TRUE,
  grid = TRUE,
  yt = "x",
  add = FALSE,
  cols = c("green", "red", "darkgreen", "orange"),
  lwds = c(3, 3),
  ltype = "l",
  lpch = 16,
  lcex = 1,
  gy = 20
)
}
\arguments{
\item{z}{an output of \code{\link{boundsPredprob}} and
\code{\link{boundsPostprob}}}

\item{area}{color the futility/efficacy stop area}

\item{grid}{show the grid of y-axis or not}

\item{yt}{indicate the y axis: response rate is "p", number of responses
is "x"}

\item{add}{add the boundary lines or not}

\item{cols}{specific the color of \verb{[1]} efficacy area \verb{[2]} futility area
\verb{[3]} efficacy boundary \verb{[4]} futility boundary}

\item{lwds}{line width (a two dimensional vector)}

\item{ltype}{line type}

\item{lpch}{a value of plotting characters or symbols}

\item{lcex}{a value giving the amount by which plotting characters and
symbols should be scaled relative to the default.}

\item{gy}{a value to adjust the gray level of the plot (default 20),
applies when \code{grid==TRUE}}
}
\value{
nothing, only produces the plot as side effect
}
\description{
This function will plot the response rate on the futility and efficacy
boundaries, and is used together with \code{\link{boundsPredprob}}
and \code{\link{boundsPostprob}}
}
\examples{
# examples
plotBounds(
  boundsPostprob(
    looks = c(10, 20, 30, 40),
    p0 = 0.20,
    tL = 0.10,
    tU = 0.90,
    parE = c(1, 1)
  ),
  yt = "p",
  add = TRUE
)
plotBounds(
  boundsPredprob(
    looks = c(10, 20, 30, 40),
    p0 = 0.20,
    tT = 0.80,
    phiL = 0.10,
    phiU = 0.90,
  ),
  yt = "x"
)
plotBounds(
  boundsPredprob(
    looks = c(10, 20, 30, 40),
    p0 = 0.20,
    tT = 0.80,
    phiL = 0.10,
    phiU = 0.90,
  ),
  yt = "p"
)
}
\keyword{graphics}
