% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecision.R
\name{plotDecision}
\alias{plotDecision}
\title{Plot a summary plot corresponding to the sumTable output}
\usage{
plotDecision(data, go_cut, stop_cut)
}
\arguments{
\item{data}{(\code{data.frame}):\cr sourced \code{\link{data.frame}} from \link{(sumTable)}}

\item{go_cut}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{stop_cut}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}
}
\value{
the \code{data} item which was imputed to the function
}
\description{
This function will return a plot showing a curve of the prob of a meaningful improvement over estunated diff
and a curve of the prob of a poor improvement over estunated diff
}
\examples{
# sumTable ----
sumTable(
  x = 10,
  n = 20,
  parX = c(1, 1),
  go_cut = 0.8,
  stop_cut = 0.4
)

summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    TotalSample = 25,
    parX = c(1, 52),
    go_cut = 0.2,
    stop_cut = 0.05
  )
)

plotDecision(summaries, go_cut = 60, stop_cut = 60)

# plotting more results
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    n = 25,
    parX = c(1, 52),
    go_cut = 0.5,
    stop_cut = 0.2
  )
)

plotDecision(summaries, go_cut = 10, stop_cut = 2)
}
\keyword{graphics}
