% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOc.R
\name{plotOc}
\alias{plotOc}
\title{Display the operating characteristics using an oc object}
\usage{
plotOc(z)
}
\arguments{
\item{z}{returned oc value}
}
\value{
nothing, only plots as side effect
}
\description{
Reads results from \code{\link{ocPredprob}}, \code{\link{ocPostprob}}
etc. and displays a bar plot of the operating characteristics
}
\examples{
# get operating character result from oc.postprob

res1 <- ocPostprob(nn = c(10, 20, 30), truep = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8, parE = c(1, 1), sim = 10000)
res1$oc

plotOc(res1)
}
\keyword{graphics}
