% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprobBeta}
\alias{postprobBeta}
\title{Posterior Probability of Efficacy Given Beta Prior}
\usage{
postprobBeta(x, n, p, a = 1, b = 1)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes.}

\item{n}{(\code{number}):\cr number of patients.}

\item{p}{(\code{number}):\cr threshold set to compute posterior probability.}

\item{a}{(\code{matrix}):\cr first parameter \code{alpha} of the beta prior (successes).}

\item{b}{(\code{matrix}):\cr second parameter \code{beta} of the beta prior (failures).}
}
\value{
The posterior probability that the response rate P_E is above a threshold p.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute the posterior probability to be above threshold assuming a beta prior
on the response rate such that  \code{Pr(P_E > p | data)}. Prior is \code{P_E ~ beta(a, b)},
with default set to be a uniform or beta(1,1).

We observed \code{x} successes in n trials and so the posterior is
\code{P_E | data  ~ beta(a + x, b + n - x)}.
}
