% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{postprobDist}
\alias{postprobDist}
\title{Compute the posterior probability with beta prior on SOC}
\usage{
postprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(1, 1),
  weights,
  parS = c(1, 1),
  weightsS
)
}
\arguments{
\item{x}{(``):\cr number of successes (in the treatment group). Note that \code{x}
can be a vector.}

\item{n}{(``):\cr number of patients (in the treatment group).}

\item{xS}{(``):\cr number of successes in the SOC group (default: 0).}

\item{nS}{(``):\cr number of patients in the SOC group (default: 0).}

\item{delta}{(``):\cr margin by which the response rate in the treatment group should
be better than in the SOC group (default: 0).}

\item{relativeDelta}{(``):\cr should the delta be relative? (not default). If this is
\code{TRUE}, then a relative delta is used. This means we want to have
response at least in delta proportion of the SOC non-responding patients.
Non-responding patients rate is 1 - P_S, and we want to have P_S + (1 - P_S)
\itemize{
\item delta response rate (at least) in the treatment. That is, we evaluate the
posterior probability Pr(P_E > P_S + (1 - P_S) * delta | data).
}}

\item{parE}{(``):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components. default is a
uniform prior.}

\item{weights}{(``):\cr the mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}

\item{parS}{(``):\cr beta parameters for the SOC group (default: uniform).}

\item{weightsS}{(``):\cr weights for the SOC group (default: uniform).}
}
\value{
the posterior probability
}
\description{
`r lifecycle::badge("experimental")

Using the approach by Thall and Simon (Biometrics, 1994), evaluate the
posterior probability of having Pr(P_E > P_S + delta | data) (but see below
for relative delta margin). Both for the new treatment E as well as for the
SOC S data might be available. However the default is that no data is
available for the SOC, corresponding to the single arm trial situation. Note
that a uniform prior is the useful default for the treatment proportion,
while in the single arm trial an informative prior on the SOC proportion is
useful.

Beta mixture prior can be specified for the treatment (\code{parE}
and \code{weights} parameters) and control proportion (\code{parS} and
\code{weightsS} parameters), see \link{postprob} for details. Note
that being able to specify a beta mixture prior also on the control
treatment is e.g. important for the futility decision making (see the
\link{oc2} code).
}
\examples{
## example similar to Lee and Liu:
postprobDist(x = 16, n = 23, parE = c(0.6, 0.4), parS = c(0.6, 0.4))

## these two should give the same result:
postprobDist(
  x = 27, n = 34,
  xS = 0, nS = 0,
  delta = 0.15,
  parE = c(1, 1),
  parS = c(50007530, 49924090)
)

postprob(x = 27, n = 34, p = 0.65, parE = c(1, 1))
## ok, almost

## try out mixtures:
## play around with the beta parameters and weights to
## get a feeling.
## Note that very extreme beta parameters do no longer increase
## the return value, because then that mixture component is too
## unlikely a posteriori
postprobDist(
  x = 16, n = 23,
  parE =
    rbind(
      c(0.6, 0.4),
      c(10, 20)
    ),
  parS =
    rbind(
      c(0.6, 0.4),
      c(10, 10)
    ),
  weightsS = c(1, 3)
)
}
