% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprob.R
\name{predprob}
\alias{predprob}
\title{Compute the predictive probability that the trial will be
successful, with a fixed response rate threshold}
\usage{
predprob(x, n, Nmax, p, thetaT, parE = c(1, 1), weights)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of patients}

\item{Nmax}{maximum number of patients at the end of the trial}

\item{p}{threshold on the response rate}

\item{thetaT}{threshold on the probability to be above p}

\item{parE}{the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components. default is a
uniform prior.}

\item{weights}{the mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}
}
\value{
The predictive probability, a numeric value. In addition, a
list called \code{table} is returned as attribute of the returned number.
}
\description{
Compute the predictive probability of trial success given current data.
Success means that at the end of the trial the posterior probability
Pr(P_E > p) >= thetaT,
where p is the fixed response rate threshold.
Then the predictive probability for success is:
pp = sum over i: Pr(Y=i|x,n)*I{Pr(P_E > p|x,Y=i)>=thetaT}
where Y is the number of future responses in the treatment group and x is
the current number of responses in the treatment group (out of n).
The prior is P_E ~ beta(a, b), default uniform which is a beta(1,1).
However, also a beta mixture prior can be specified.
}
\details{
A table with the following contents will be included in the \code{table} attribute of
the return value:
i: Y=i (number of future successes in Nmax-n subjects)
py: Pr(Y=i|x) using beta-(mixture)-binomial distribution
b: Pr(P_E > p | x, Y=i) using beta (mixture) posterior
bgttheta: indicator I(b>thetaT)
}
\examples{
## The original Lee and Liu (Table 1) example: 
## Nmax=40, x=16, n=23, beta(0.6,0.4) prior distribution,
## thetaT=0.9. The control response rate is 60\%:
predprob(x=16, n=23, Nmax=40, p=0.6, thetaT=0.9, 
         parE=c(0.6, 0.4))
## So the predictive probability is 56.6\%. 12 responses
## in the remaining 17 patients are required for success.

## Lowering/Increasing the probability threshold thetaT of course increases
## /decreases the predictive probability of success, respectively:
predprob(x=16, n=23, Nmax=40, p=0.6, thetaT=0.8, 
         parE=c(0.6, 0.4))
## 70.8\%
predprob(x=16, n=23, Nmax=40, p=0.6, thetaT=0.95, 
         parE=c(0.6, 0.4))
## 40.7\%

## Instead of a fixed beta prior on the response rate, dynamic borrowing
## from a previous data set is possible by using a beta-mixture prior.
## For example, assume we have a previous trial where we saw 25 responses
## in 40 patients. We would like to use information worth 10 patients in our
## trial, but be robust against deviations from that response rate (maybe 
## because it was conducted in slightly different disease or patients).
## Then we can use a beta mixture prior, with the informative component getting
## weight of 1/4:
predprob(x=20, n=23, Nmax=40, p=0.6, thetaT=0.9, 
         parE=rbind(c(1, 1), c(25, 15)),
         weights=c(3, 1))
## Since the response rate in the historical dataset is lower (62.5\% < 69.6\%)
## than in the trial, but similar, we notice that the predictive probability
## of success is now lower.
}
\references{
Lee, J. J., & Liu, D. D. (2008). A predictive probability
design for phase II cancer clinical trials. Clinical Trials, 5(2),
93-106. doi:10.1177/1740774508089279
}
