% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprobDist.R
\name{predprobDist}
\alias{predprobDist}
\title{Compute the predictive probability that the trial will be
successful, with a prior distribution on the SOC}
\usage{
predprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  Nmax,
  NmaxControl = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(a = 1, b = 1),
  weights,
  parS = c(a = 1, b = 1),
  weightsS,
  thetaT
)
}
\arguments{
\item{x}{number of successes (in the treatment group)}

\item{n}{number of patients (in the treatment group)}

\item{xS}{number of successes in the SOC group (default: 0)}

\item{nS}{number of patients in the SOC group (default: 0)}

\item{Nmax}{maximum number of patients at the end of the trial (in the
treatment group)}

\item{NmaxControl}{maximum number of patients at the end of the trial in the
SOC group (default: 0)}

\item{delta}{margin by which the response rate in the treatment group should
be better than in the SOC group (default: 0)}

\item{relativeDelta}{see \code{\link{postprobDist}}}

\item{parE}{the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components. default is a
uniform prior.}

\item{weights}{the mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}

\item{parS}{beta prior parameters in the SOC group (default: uniform)}

\item{weightsS}{weights for the SOC group (default: uniform)}

\item{thetaT}{threshold on the probability to be used}
}
\value{
The predictive probability, a numeric value. In addition, a
list called \code{tables} is returned as attribute of the returned number.
}
\description{
Compute the predictive probability of trial success given current data.
Success means that at the end of the trial the posterior probability
Pr(P_E > P_S + delta_0 | data) >= thetaT. Then the
predictive probability for success is:
pp = sum over i: Pr(Y=i|x,n)*I{Pr(P_E > P_S + delta|x,Y=i)>=thetaT},
where Y is the number of future responses in the treatment group and x is
the current number of responses in the treatment group (out of n).
Prior is P_E ~ beta(a, b), default uniform which is a beta(1,1).
However, also a beta mixture prior can be specified. Analogously
for P_S either a classic beta prior or a beta mixture prior can be
specified.
}
\details{
Also data on the SOC might be available. Then the predictive probability is
more generally defined as
pp = sum over i, j: Pr(Y=i|x,n)*Pr(Z=j|xS, nS)*I{Pr(P_E > P_S +
delta|x,xS,Y=i,Z=j)>=thetaT}
where Z is the future number of responses in the SOC group, and xS is the
current number of responses in the SOC group.

A table with the following contents will be
included in the \code{tables} attribute of the return value
(in the case that NmaxControl is zero):
i: Y=i (number of future successes in Nmax-n subjects)
py: Pr(Y=i|x) using beta-(mixture)-binomial distribution
b: Pr(P_E > P_S + delta | x, Y=i)
bgttheta: indicator I(b>thetaT)

If NmaxControl is not zero, i.e., when data on the control treatment
is available in this trial, then a list with will be included with the
following elements:
pyz: matrix with the probabilities Pr(Y=i, Z=j | x, xS)
b: matrix with Pr(P_E > P_S + delta | x, xS, Y=i, Z=j)
bgttheta: matrix of indicators I(b>thetaT)
}
\examples{
## The original Lee and Liu (Table 1) example: 
## Nmax=40, x=16, n=23, beta(0.6,0.4) prior distribution,
## thetaT=0.9. The control response rate is 60\%:
predprob(x=16, n=23, Nmax=40, p=0.6, thetaT=0.9, 
         parE=c(0.6, 0.4))
## So the predictive probability is 56.6\%. 12 responses
## in the remaining 17 patients are required for success.

## Now to the modified example, where the control response 
## rate p has a beta(7, 11) distribution.
## For example if the 60\% threshold stems from a trial
## with 10 patients, of which 6 were observed with response.
predprobDist(x=16, n=23, Nmax=40, delta=0, thetaT=0.9, 
             parE=c(0.6, 0.4), parS=c(7, 11))
## Now only 11 responses are needed in the remaining 17 patients
## in order to have a successful trial. 
## Also, the predictive probability for a Go decision is now
## 70.8\% instead of merely 56.6\%.

## If the response threshold of 60\% stems from an SOC estimate (e.g. 50\% = 5/10)
## and a margin of e.g. 10\%, then the margin can be specified with
## the delta argument:
predprobDist(x=16, n=23, Nmax=40, delta=0.1, thetaT=0.9, 
             parE=c(1, 1), parS=c(6, 11))
## This again changes the result: Only 10 future responses are required,
## the predictive probability of success is now 80\%.

## Next we will use a beta mixture prior, with a weight of 1/3 for an
## informative beta(50, 10) prior, on the experimental response rate:
predprobDist(x=16, n=23, Nmax=40, delta=0.1, thetaT=0.9,
             parE=rbind(c(1, 1), c(50, 10)),
             weights=c(2, 1), parS=c(6, 11))
## Now still 10 future responses are required for success, but the predictive
## success probability is higher (87.2\%).

## We can also have additional controls in our trial. For example
## assume that in the interim analysis we had 10 control patients observed
## with 5 responses out of 20 in total at the end. 
## We also had historical control data (say 60 patients of
## which 20 responded), but in order to be robust against changes in the response 
# rate over time, we will again use a beta mixture, putting only 1/3 of weight
## on the historical controls:
predprobDist(x=16, n=23, xS=5, nS=10,
             Nmax=40, NmaxControl=20, 
             delta=0.1, thetaT=0.9,
             parE=rbind(c(1, 1), c(50, 10)),
             weights=c(2, 1), 
             parS=rbind(c(1, 1), c(20, 40)),
             weightsS=c(2, 1))
## Now we only have 59.9\% predictive probability of success. The bgttheta matrix
## lists the cases where we would have success, for all possible combinations
## of future responses in the control and experimental arms. We see that the 
## success threshold for experimental responses depends on the number of 
## control responses.

}
\references{
Lee, J. J., & Liu, D. D. (2008). A predictive probability
design for phase II cancer clinical trials. Clinical Trials, 5(2),
93-106. doi:10.1177/1740774508089279
}
