% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{qbetaMix}
\alias{qbetaMix}
\title{Beta-mixture quantile function}
\usage{
qbetaMix(q, par, weights, lower.tail)
}
\arguments{
\item{q}{(\code{numeric}):\cr the required quantile.}

\item{par}{(\code{number}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}

\item{weights}{(\code{matrix}):\cr the mixture weights of the beta mixture prior.}
}
\value{
the abscissa.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Note that \code{x} can be a vector.
}
