% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTable.R
\name{sumTable}
\alias{sumTable}
\title{Return summary statistics under one response scenario}
\usage{
sumTable(x, n, go_cut, stop_cut, parX, parY = c(0.5, 0.5), Round = 2)
}
\arguments{
\item{x}{(\code{number}):\cr number of responses}

\item{n}{(\code{number}):\cr sample size}

\item{go_cut}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{stop_cut}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}

\item{parY}{(\code{numeric}):\cr two parameters of \code{Y}'s Beta distribution (Treatment)}

\item{Round}{(\code{number}):\cr Digit rounding of the output statistics}

\item{treat_par}{(\code{numeric}):\cr non-negative parameters of the beta prior of the treatment, default Beta(0.5,0.5)}
}
\value{
A vector with the results.
}
\description{
This function will calculate the summary statistics for a specific response
outcome scenario.
}
\examples{
sumTable(x = 2,
         n = 25,
         parX = c(1, 52),
         parY = c(1, 1),
         go_cut = 0.2,
         stop_cut = 0.05)

# for multiple response scenarios (e.g. 0 to 8 responses out of 25)
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    TotalSample = 25,
    parX = c(1, 52),
    go_cut = 0.2,
    stop_cut = 0.05
  )
)
summaries


#
# > sumTable(2, 25, parX = c(1, 52), go_cut = 0.2, stop_cut = 0.05, YPri = c(1, 1))
# summaries
# # resp             2.00
# obs ORR [\%]        8.00
# mode [\%]           6.48
# CI lower [\%]       0.71
# CI upper [\%]      20.68
# prob.go [\%]        5.87
# prob.nogo [\%]     26.59
}
