% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TTEdata}
\alias{TTEdata}
\title{Example of time to event dataset}
\format{
A data frame with 80 obs. of  12 variables:
\describe{
\item{ID}{Patient number formatted as USUBJID}
\item{treatment}{Factor w/ 2 levels: "Control" and "Experimental"}
\item{PFStime}{progression free surivival, days}
\item{PFSevent}{censoring indicator for PFS}
\item{DoR}{duration of response}
\item{responder}{responder status, used as censoring variable for DoR}
\item{TtR}{time to response}
\item{TtRevent}{censoring indicator for TtR}
\item{OStime}{survival time, days}
\item{OSevent}{death indicator}
\item{SEX}{covariate}
\item{AGE}{covariate}
}
}
\usage{
TTEdata
}
\description{
An example dataset containing oncology time to event endpoints
formatted for time to event analysis functions in the package
}
\keyword{datasets}
