% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOc.R
\name{vintage_plotOc}
\alias{vintage_plotOc}
\title{Display the operating characteristics using an oc object}
\usage{
vintage_plotOc(oc)
}
\arguments{
\item{oc}{(\code{list}):\cr returned oc parameters}
}
\value{
nothing, only plots as side effect
}
\description{
Reads results from \code{\link[=ocPredprob]{ocPredprob()}}
etc. and displays a bar plot of the operating characteristics
}
\examples{
# get operating character result from oc.postprob

oc <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.4, p0 = 0.2,
  p1 = 0.3, tL = 0.6, tU = 0.8, parE = c(1, 1), sim = 100, wiggle = FALSE
)
oc$oc
# plotOc(oc)


oc <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.4, p0 = 0.2,
  p1 = 0.3, tL = 0.6, tU = 0.8, parE = c(1, 1), sim = 100, wiggle = TRUE
)
oc$oc
}
\keyword{graphics}
