% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiff.R
\name{betadiff}
\alias{betadiff}
\alias{dbetadiff}
\alias{pbetadiff}
\alias{qbetadiff}
\title{The Difference between Two Beta Distributions}
\usage{
dbetadiff(z, parY, parX)

pbetadiff(q, parY, parX)

qbetadiff(p, parY, parX)
}
\arguments{
\item{z}{vector of differences}

\item{parY}{two parameters of Y's beta distribution (Treatment)}

\item{parX}{two parameters of X's beta distribution (Control)}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}
}
\value{
\code{dbetadiff} gives the density, \code{pbetadiff} the distribution function,
\code{qbetadiff} the quantile function.
}
\description{
Density, distribution function and quantile function for
the distribution of the difference of two beta distributions with parameters parX and parY
}
\examples{

## Plot pdf of the betadiff distribution
parX <- c(1, 52)
parY <- c(5.5, 20.5)

z <- seq(from=-1, to=1, length=100)
plot(z, dbetadiff(z, parY=parY, parX=parX),
     type="l")

## calculate probability

## for Go:
test <- integrate(f=dbetadiff,
                  parY=parY,
                  parX=parX,
                  lower=0.15,
                  upper=1)
str(test)
test$value

## vary close to the Monte Carlo result:
mean(rbeta(n=1e6, parY[1], parY[2]) -
     rbeta(n=1e6, parX[1], parX[2]) > 0.15)


## similar case for Stop:
integrate(f=dbetadiff,
          parY=parY,
          parX=parX,
          lower=-1,
          upper=0.5)

mean(rbeta(n=1e6, parY[1], parY[2]) -
     rbeta(n=1e6, parX[1], parX[2]) < 0.5)

test <- qbetadiff(p=0.2,
                  parY=parY,
                  parX=parX)
pbetadiff(q=test,
          parY=parY,
          parX=parX)
}
