% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiff.R
\name{betadiff}
\alias{betadiff}
\alias{dbetadiff}
\title{The Difference between Two Beta Distributions}
\usage{
dbetadiff(z, parY, parX)
}
\arguments{
\item{z}{(\code{numeric}):\cr vector of differences between Control and Treatment arms such that \code{Z = Y-X}}

\item{parY}{(\code{numeric}):\cr two parameters of \code{Y}'s Beta distribution (Treatment)}

\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}
}
\value{
\code{dbetadiff} gives the density
}
\description{
Density, distribution function and quantile function for
the distribution of the difference of two Beta distributions with parameters parX and parY.
We denote \code{X} and \code{Y} as two random variables representing the response rate of Control and Treatment
group respectively. The assignment of Control and Treatment is practically interchangeable.
We denote \code{Z} as the difference between two groups such that \code{Z = Y-X}.
}
\note{
\code{X} and \code{Y} can be either Control or Treatment and \code{Z = X-Y}, subject to assumptions
}
\examples{
# dbetadiff ----

# We calculate and plot of density (pdf) of the Beta distribution of the difference.
parX <- c(1, 52)
parY <- c(5.5, 20.5)

# The difference between Control and Treatment is denoted as z.
z <- seq(from = -1, to = 1, length = 100)
plot(z, dbetadiff(z, parY = parY, parX = parX),
  type = "l"
)

# dbetadiff ----
# Calculate probability of Go, if difference was at least 15\%.
test <- integrate(
  f = dbetadiff,
  parY = parY,
  parX = parX,
  lower = 0.15,
  upper = 1
)
str(test)
test$value

# Calculate probability of Stop, if difference was at most 50\%.
integrate(
  f = dbetadiff,
  parY = parY,
  parX = parX,
  lower = -1,
  upper = 0.5
)
}
