% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsPostprob.R
\name{boundsPostprob}
\alias{boundsPostprob}
\title{Decision cutpoints for boundary (based on posterior probability)}
\usage{
boundsPostprob(nvec, p0, p1 = p0, tL, tU, a, b)
}
\arguments{
\item{nvec}{a vector of number of patients}

\item{p0}{the efficacy threshold parameter in the postprob function}

\item{p1}{the futility threshold parameter in the postprob function
(default = p0)}

\item{tL}{futility boundary probability threshold}

\item{tU}{efficacy boundary probability threshold}

\item{a}{the alpha parameter of the beta prior of treatment group}

\item{b}{the beta parameter of the beta prior of treatment group}
}
\value{
A matrix where for each sample size in \code{nvec}, this function
returns the maximum number of responses that meet the futility
threshold (xL), its corresponding response rate (pL), posterior probability
(postL), upper bound of one sided 95\% CI for the response rate based on an
exact binomial test (UciL), and the same boundary parameters for efficacy:
the minimal number of responses that meet the efficacy threshold (xU),
the corresponding response rate (pU), posterior probability (postU) and
the lower bound of one sided 95\% CI for the response rate based on exact
binomial test (LciU).
}
\description{
This function is used to identify the efficacy and futility
boundaries based on posterior probabilities, i.e.:
Efficacy boundary: find minimum x (xU) where Pr(P>p0|x,n,a,b) >= tU and
Futility boundary: find maximum x (xL) where Pr(P>p1|x,n,a,b) <= tL
}
\examples{
## 40 pts trial with interim looks after each 10 pts.,
## efficacy decision if more than 90\% probability to be above 20\% ORR,
## futility decision if less than 10\% probability to be above 20\% ORR,
## with uniform prior (i.e. beta(1, 1)) on the ORR:
boundsPostprob(nvec=c(10,20,30,40), p0=0.20,
               tL=0.10, tU=0.90, a=1, b=1)
## From this we see e.g. that at the third IA at 30 pts, we would stop for futility
## if 5 or less patients responded, and for efficacy if 9 or more pts responded.
}
\keyword{graphics}
