% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsPostprob.R
\name{boundsPostprob}
\alias{boundsPostprob}
\title{Decision cutpoints for boundary (based on posterior probability)}
\usage{
boundsPostprob(looks, p0, p1 = p0, tL, tU, parE = c(1, 1), weights)
}
\arguments{
\item{looks}{(\code{numeric}):\cr A vector of number of patients in each look.}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{p1}{(\code{number}):\cr upper Efficacy threshold of response rate.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{parE}{(\code{matrix}):\cr the beta parameters matrix, with \code{K} rows and 2 columns,
corresponding to the beta parameters of the \code{K} components.}

\item{weights}{(\code{vector}):\cr The mixture weights of the beta mixture prior.}
}
\value{
A matrix for each same size in \code{looks}. For each sample size, the following is returned:
\itemize{
\item \code{xL} : the maximum number of responses that meet the futility threshold.
\item \code{pL} : response rate corresponding to \code{xL}.
\item \code{postL}: posterior probability corresponding to \code{xL}, i.e. Pr(RR < p0 | xL, n, a, b).
\item \code{pL_upper_ci} : upper bound of one sided 95\% CI for the response rate \code{pL} based on an
exact binomial test.
\item \code{xU} : the minimal number of responses that meet the efficacy threshold.
\item \code{pU} : response rate corresponding to \code{xU}.
\item \code{postU} : posterior probability corresponding to \code{xU}, i.e. Pr(RR > p1 |xU, n, a, b).
\item \code{pU_lower_ci} : lower bound of one sided 95\% CI for the response rate \code{pU} based on exact
binomial test.
}
}
\description{
This function is used to identify the efficacy and futility
boundaries based on the following rules:
Efficacy boundary: find minimum x (xU) where Pr(RR > p1 | x, n, a, b) >= tU and
Futility boundary: find maximum x (xL) where Pr(RR < p0 | x, n, a, b) >= tL
}
\examples{
# 40 pts trial with interim looks after each 10 pts.,
# Efficacy decision if more than 80\% probability to be above 20\% ORR,
# Futility decision if more than 60\% probability to be below 15\% ORR,
# with uniform prior (i.e. beta(1, 1)) on the ORR:
boundsPostprob(
  looks = c(10, 20, 30, 40),
  p0 = 0.15,
  p1 = 0.20,
  tL = 0.60,
  tU = 0.80,
  parE = c(1, 1)
)

# 40 pts trial with interim looks at 7 and 20 pts.
# Efficacy decision if more than 80\% probability to be above 20\% ORR,
# Futility decision if more than 60\% probability to be below 15\% ORR,
# with mixed prior and weights:
boundsPostprob(
  looks = c(7, 20, 40),
  p0 = 0.15,
  p1 = 0.20,
  tL = 0.60,
  tU = 0.80,
  parE = rbind(c(1, 19), c(2, 10)),
  weights = c(0.2, 0.8)
)
}
