% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsPredprob.R
\name{boundsPredprob}
\alias{boundsPredprob}
\title{Decision cutpoints for boundary (based on predictive probability)}
\usage{
boundsPredprob(nvec, Nmax = max(nvec), p, tT, phiL, phiU, a, b)
}
\arguments{
\item{nvec}{a vector of number of patients}

\item{Nmax}{maximum number of patients at the end of the trial
(default: maximum of \code{nvec})}

\item{p}{threshold on the response rate}

\item{tT}{threshold on the posterior probability to be above p}

\item{phiL}{futility boundary predictive probability threshold}

\item{phiU}{efficacy boundary predictive probability threshold}

\item{a}{the alpha parameter of a beta prior of treatment group}

\item{b}{the beta parameter of a beta prior of treatment group}
}
\value{
A matrix where for each sample size in \code{nvec}, this function
returns the maximum number of responses that meet the futility
threshold (xL), its corresponding response rate (pL), predictive probability
(ppL) and posterior probability (postL), the upper bound of one
sided 95\% CI for the response rate based on an
exact binomial test (UciL), and the same boundary parameters for efficacy:
the minimal number of responses that meet the efficacy threshold (xU),
the corresponding response rate (pU), predictive probability
(ppL) and posterior probability (postU), the lower bound of one sided
95\% CI for the response rate based on exact binomial test (LciU).
}
\description{
This function is used to identify the efficacy boundary and futility
boundary based on predictive probabilities, i.e.:
Efficacy boundary: find minimum x (xU) where
Pr(Pr(P > p | x, Y) >= tT | x) > phiU,
Futility boundary: find maximum x (xL) where
Pr(Pr(P > p | x, Y) >= tT | x) < phiL
}
\examples{
## 40 pts trial with interim looks after each 10 pts.,
## final efficacy decision if more than 80\% probability to be above 20\% ORR,
## final futility decision otherwise.
## Interim efficacy decision if more than 90\% predictive probability reach this,
## interim futility decision if less than 10\% predictive probability.
## Uniform prior (i.e. beta(1, 1)) on the ORR:
boundsPredprob(
  nvec = c(10, 20, 30, 40), p = 0.20, tT = 0.80,
  phiL = 0.10, phiU = 0.90, a = 1, b = 1
)
## From this we see e.g. that at the first IA at 10 pts, we would stop for futility
## if no patient responded, and for efficacy if 4 or more pts responded.
}
\keyword{graphics}
