% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-binomial density function}
\usage{
dbetabinom(x, m, a, b, log = FALSE)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes.}

\item{m}{(\code{number}):\cr number of trials.}

\item{a}{(\code{numeric}):\cr first parameter of the beta distribution.}

\item{b}{(\code{numeric}):\cr second parameter of the beta distribution.}

\item{log}{(\code{flag}):\cr whether to return the log density value (not default).}
}
\value{
The density values of the beta-binomial distribution at \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the density function of the beta-binomial distribution.

The beta-binomial density function has the following form:
\verb{p(x) = (m! / (x!*(m-x)!)) * Beta(x+a,m-x+b) / Beta(a,b)}
}
\note{
\code{x}, \code{a} and \code{b} can be vectors.
}
\examples{
dbetabinom(x = 2, m = 29, a = 0.2, b = 0.4, log = FALSE)

# Can also specify x as a vector.

dbetabinom(x = 1:28, m = 29, a = 0.2, b = 0.4, log = FALSE)
}
