% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-binomial density function}
\usage{
dbetabinom(x, m, a, b)
}
\arguments{
\item{x}{number of successes}

\item{m}{number of trials}

\item{a}{first parameter of the beta distribution}

\item{b}{second parameter of the beta distribution}
}
\value{
the density values of the beta-binomial distribution at \code{x}
}
\description{
Calculates the density function of the beta-binomial distribution
}
\details{
Note that \code{x} can be a vector.

The beta-binomial density function has the following form:
\deqn{p(x) = (m! / (x!*(m-x)!)) * Beta(x+a,m-x+b) / Beta(a,b)}
}
\examples{
## Calculating the beta binomial density, x = 2; m = 29; a = 0.2; b = 0.4
##
## when p(x) = choose(29,2)*beta(2.2,27.4)/beta(0.2,0.4) = 0.04286893
##
dbetabinom(2, 29, 0.2, 0.4)

## Can also specify x as a vector, x = 1:28
##
##
dbetabinom(1:28, 29, 0.2, 0.4)
}
