% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-binomial density function}
\usage{
dbetabinom(x, m, a, b)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes.}

\item{m}{(\code{number}):\cr number of trials.}

\item{a}{(\code{number}):\cr first parameter of the beta distribution.}

\item{b}{(\code{number}):\cr second parameter of the beta distribution.}
}
\value{
The density values of the beta-binomial distribution at \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the density function of the beta-binomial distribution

Note that \code{x} can be a vector.

The beta-binomial density function has the following form:
\verb{p(x) = (m! / (x!*(m-x)!)) * Beta(x+a,m-x+b) / Beta(a,b)}
}
\examples{
dbetabinom(x = 2, m = 29, a = 0.2, b = 0.4)

# Can also specify x as a vector.
dbetabinom(x = 1:28, m = 29, a = 0.2, b = 0.4)
}
