% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{dbetaMix}
\alias{dbetaMix}
\title{Beta-Mixture density function}
\usage{
dbetaMix(x, par, weights, log = FALSE)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes.}

\item{par}{(\code{matrix}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior of length K.
Each element corresponds to the row of beta parameters in \code{par}.}

\item{log}{(\code{flag}):\cr whether log values of the beta-mixture density function are returned.}
}
\value{
the (log) density values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The beta-mixture density can be calculated with the combination of K set of beta parameters and K
length of weights.
}
\details{
Calculating beta-mixture density values of support \code{x}.
}
