% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{h_get_bounds}
\alias{h_get_bounds}
\title{Generating bounds for the Integration of Beta Mixture Posterior}
\usage{
h_get_bounds(controlBetamixPost)
}
\arguments{
\item{betamixPost}{(\code{list}):\cr arguments of \code{par}and \code{weights} of Beta Mixture Posterior in format list. See \verb{[getBetaMix()]}.}

\item{par}{(\code{matrix}):\cr the beta parameters matrix, with \code{K} rows and 2 columns,
corresponding to the beta parameters of the \code{K} components.}

\item{weights}{(\code{vector}):\cr The mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}
}
\value{
Integrand function
}
\description{
Using the quantile of the Beta Mixture Distribution from parameters given by standard of care \code{SOC} or
experimental group \code{E} to determine bounds as inputs to \verb{[stats::integrate()]}
}
\keyword{internal}
