% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{h_get_decision}
\alias{h_get_decision}
\title{Generating random decision and sample size looks}
\usage{
h_get_decision(nnr, truep, p0, p1, parE = c(1, 1), nnE, nnF, tL, tU, orig_nnr)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE}and \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate.}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{p1}{(\code{number}):\cr upper Efficacy threshold of response rate.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{orig_nnr}{(\code{numeric}):\cr unique values of original, user-input values of \code{nnE} and \code{nnF}, defined by user-facing
function and aid to \code{plotOc()}'s x axis input when \code{wiggle = TRUE}.}
}
\value{
A list with the following elements :
\itemize{
\item \code{decision} : decision \code{flag} with \code{TRUE} for Go, \code{FALSE} for Stop, \code{NA} for Gray zone.
\item \code{all_sizes} : resulting number of look size, anything below maximum
look size is an indicated interim, Futility or Efficacy or both.
}
}
\description{
A helper function for \code{\link[=ocPostprob]{ocPostprob()}} to generate numeric of decisions \code{decisions} and random looks \code{all_sizes}.
}
\keyword{internal}
