% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprobDist.R
\name{h_get_decisionDist}
\alias{h_get_decisionDist}
\title{Evaluate posteriors based on  Efficacy and Futility thresholds in two-armed trials.}
\usage{
h_get_decisionDist(
  nnr,
  nnrE,
  nnrF,
  truep,
  parE = c(1, 1),
  parS = c(1, 1),
  tL,
  tU,
  deltaF,
  deltaE,
  relativeDelta
)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr the union of \code{nnE}and \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{parE}{(``):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(``):\cr alpha and beta parameters for the prior on the control population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{deltaF}{(``):\cr delta for efficacy: \code{P(P_E > P_S + deltaE)} should be large
to stop for efficacy.}

\item{deltaE}{(``):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients
exists and is used if \code{TRUE}.
See \code{\link[=postprobDist]{postprobDist()}}.}

\item{<parse error>}{(\if{html}{\out{<parse error>}}):\cr \if{html}{\out{<parse error>}}}
}
\value{
A list of the following objects :
\itemize{
\item \code{decision} : resulting numeric of decision, one of \code{TRUE} for Go, \code{FALSE} for Stop, \code{NA} for Gray zone.
\item \code{all_sizes} : resulting numeric of look size, anything below maximum
look size is an indicated interim, Futility or Efficacy or both.
}
}
\description{
The helper function to evaluate generated posteriors and create an "array"
or "vector" of decision results based on user defined thresholds.
\code{h_get_decisionDist} was adapted by incorporating \code{postprobDist()}
instead of \code{postprob()} to evaluate this posterior.
The methodology is from Thall & Simon 1994 (Biometrics).
}
\keyword{internal}
