% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprobDist.R
\name{h_get_decisionDist}
\alias{h_get_decisionDist}
\title{Evaluate Posteriors based on Efficacy and Futility Thresholds in Two-Armed trials}
\usage{
h_get_decisionDist(
  nnr,
  nnrE,
  nnrF,
  truep,
  parE = c(1, 1),
  parS = c(1, 1),
  tL,
  tU,
  deltaE,
  deltaF,
  relativeDelta
)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE}and \code{nnF}.}

\item{nnrE}{(\code{numeric}):\cr looks with random distance, if applied on \code{nnE}.}

\item{nnrF}{(\code{numeric}):\cr looks with random distance, if applied on \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{S} group. See details.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{deltaE}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Efficacy looks only.
Note that this can also be negative as well.}

\item{deltaF}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Futility looks only.
Note that this can also be negative as well.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients
is included as the margin between treatment and control group. See note.}
}
\description{
A helper function adapted from \code{\link[=h_get_decision]{h_get_decision()}} by
replacing the internal use of \code{\link[=postprob]{postprob()}} with \code{\link[=postprobDist]{postprobDist()}}
to generate the \code{decision} and random looks \code{all_sizes}.
}
\keyword{internal}
