% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocRctPostprobDist.R
\name{h_get_decisionDist_rct}
\alias{h_get_decisionDist_rct}
\title{Generating random decision and sample size looks}
\usage{
h_get_decisionDist_rct(
  nnr,
  nnrE,
  nnrF,
  pE,
  pS,
  parE = c(1, 1),
  parS = c(1, 1),
  tL,
  tU,
  deltaE,
  deltaF,
  relativeDelta,
  randRatio = 1,
  Nmax
)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE} and \code{nnF}.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{randRatio}{(\code{numeric}):\cr The randomisation ratio between treatment and control. Must be greater than 0 and maximum of 1.}

\item{Nmax}{(\code{number}):\cr The max sample size or the sample size of final look.}
}
\value{
A list with the following elements :
\itemize{
\item \code{decision} : decision \code{flag} with \code{TRUE} for Go, \code{FALSE} for Stop, \code{NA} for Gray zone.
\item \code{all_sizes} : resulting number of look size, anything below maximum
look size is an indicated interim, Futility or Efficacy or both.
\item \code{nActive} : mean of look size for Active arm.
\item \code{nControl} : mean of look size for Control arm.
}
}
\description{
A helper function for \code{\link[=ocRctPostprobDist]{ocRctPostprobDist()}} to generate numeric of decisions \code{decisions} and
random looks \code{all_sizes}.
}
\keyword{internal}
