% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{h_get_distance}
\alias{h_get_distance}
\title{Generating random distance in given looks for sample sizes for efficacy and futility.}
\usage{
h_get_distance(nn)
}
\arguments{
\item{nn}{: number or numeric
the union of \code{nnE} and \code{nnF} (if futility analysis or looks exists) supplied}
}
\value{
A numeric with \code{length(nn)-1} elements.
}
\description{
A helper function for \code{ocPostprob} to generate random distance's wiggle room around looks \code{nn}.
Numeric looks \code{nn} must be of minimum two elements and will generate \code{length(nn)-1} distances.
}
\examples{
examples / ocPostprob.R
}
\keyword{internal}
