% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{h_get_looks}
\alias{h_get_looks}
\title{Generating looks from random distance}
\usage{
h_get_looks(dist, nnE, nnF)
}
\arguments{
\item{dist}{(\verb{numeric or logical}):\cr distance for random looks around the look locations in \code{nn},
where \code{dist} is generated from \code{\link[=h_get_distance]{h_get_distance()}} in a numeric of at least one element.
If \code{NULL}, only one location look will be set at \code{nnE} or \code{nnF}.}

\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}
}
\value{
Uses distance from \code{\link[=h_get_distance]{h_get_distance()}} to add to looks, creating wiggled looks:
\itemize{
\item \code{nnrE} is the result for Efficacy looks with random distance added.
\item \code{nnrF} is the result for Futility looks with random distance added.
}
}
\description{
A helper function for \code{\link[=ocPostprob]{ocPostprob()}} that applies the numeric element of \code{dist} to looks \code{nn}.
}
\keyword{internal}
