% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{h_get_oc_predprob}
\alias{h_get_oc_predprob}
\title{Generating random decision and sample size looks}
\usage{
h_get_oc_predprob(all_sizes, nnr, decision, nnrE, nnrF, Nmax)
}
\arguments{
\item{nnr}{(\code{numeric}):\cr union of \code{nnE} and \code{nnF}.}

\item{<parse error>}{(\if{html}{\out{<parse error>}}):\cr \if{html}{\out{<parse error>}}}
}
\value{
A list of results containing :
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\description{
A helper function for \code{\link[=ocPredprob]{ocPredprob()}} to generate numeric of decisions \code{decisions} and
random looks \code{all_sizes} for \code{decision1 == TRUE}.
}
\keyword{internal}
