% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{h_get_oc_predprob}
\alias{h_get_oc_predprob}
\title{Generating random decision and sample size looks}
\usage{
h_get_oc_predprob(all_sizes, nnr, decision, nnrE, nnrF, Nmax = Nmax)
}
\arguments{
\item{all_sizes}{(\code{numeric}):\cr Sample sizes of all simulated \code{length(sim)} times if \code{dist} applied.}

\item{nnr}{(\code{numeric}):\cr union of \code{nnE} and \code{nnF}.}

\item{decision}{(\code{numeric}):\cr Go, Stop or Gray Zine decisions of all looks simulated \code{length(sim)} times.}

\item{nnrE}{(\code{numeric}):\cr Looks with random distance, if applied on \code{nnE}}

\item{nnrF}{(\code{numeric}):\cr Looks with random distance, if applied on \code{nnF}}

\item{Nmax}{(\code{number}):\cr Maximum sample size or final look from the maximum of \code{union(c(nnF, nnE))}}

\item{sim}{(\code{number}):\cr \code{length(all_sizes)} iterations for frequentist calculation of operating characteristics.}
}
\value{
A list of results containing :
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\description{
A helper function for \code{\link[=ocPredprob]{ocPredprob()}} to generate numeric of decisions \code{decisions} and
random looks \code{all_sizes} for \code{decision1 == TRUE}.
}
\keyword{internal}
