% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{h_getBetamixPost}
\alias{h_getBetamixPost}
\title{Compute Beta-Mixture-Binomial Posterior Distribution's parameters and weights.}
\usage{
h_getBetamixPost(x, n, par, weights)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes. Can be a vector of \code{length(x) > 1}.}

\item{par}{(\code{matrix}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior of length K.
Each element of the numeric is assigned to a pair of beta parameters.}
}
\value{
A list with the updated beta parameters and weights
}
\description{
A helper function that computes the posterior parameters of a beta-mixture-binomial distribution.
}
\keyword{internal}
