% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{h_integrand}
\alias{h_integrand}
\title{The Posterior Beta Mixture Integrand when Delta is absolute.}
\usage{
h_integrand(p_s, delta, x, activeBetamixPost, controlBetamixPost)
}
\value{
An R function that is an argument for \verb{[stats::integrate()]}.
}
\description{
The helper function to generate Integrand function when relative Delta not given.
A numerical integration to compute this probability is given on p.338
}
\keyword{internal}
