% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{h_integrand_relDelta}
\alias{h_integrand_relDelta}
\title{The Posterior Beta Mixture Integrand when \code{delta} is relative}
\usage{
h_integrand_relDelta(p_s, delta, activeBetamixPost, controlBetamixPost)
}
\arguments{
\item{p_s}{(\code{number}):\cr probability of success or response rate of standard of care or control or \code{S} group.}

\item{delta}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group.
Note that this can also be negative, e.g. when non-inferiority is being assessed.}

\item{activeBetamixPost}{(\code{list}):\cr a list of posterior parameters of a beta-mixture-binomial distribution with generic names
\code{par} and \code{weights}. See \verb{[getBetaMix()]}.}

\item{controlBetamixPost}{(\code{list}):\cr a list of posterior parameters of a beta-mixture-binomial distribution with generic names
\code{par} and \code{weights}. See \verb{[getBetaMix()]}.}
}
\value{
Function that is an argument for \verb{[stats::integrate()]}.
}
\description{
The helper function to generate Integrand function when \code{relativeDelta = TRUE}.
}
\keyword{internal}
