% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprobDist.R
\name{h_predprobdist}
\alias{h_predprobdist}
\title{The predictive probability of success in two-arm studies.}
\usage{
h_predprobdist(
  x,
  n,
  xS,
  nS,
  Nmax,
  NmaxControl,
  delta,
  relativeDelta,
  parE,
  parS,
  weights,
  weightsS,
  thetaT
)
}
\arguments{
\item{x}{(\code{number}):\cr number of successes in the \code{E} group at interim.}

\item{n}{(\code{number}):\cr number of patients in the \code{E} group at interim.}

\item{xS}{(\code{number}):\cr number of successes in the \code{E} group at interim.}

\item{Nmax}{(\code{number}):\cr maximum number of patients in the \code{E} group at final analysis.}

\item{NmaxControl}{(\code{number}):\cr maximum number of patients in the \code{E} group at final analysis.}
}
\value{
A \code{list} is returned with names \code{result} for predictive probability and
\code{table} of numeric values with counts of responses in the remaining patients, \code{density} for
probabilities of these counts, \code{posterior} for corresponding probabilities to be above threshold,
and \code{success}for trial success indicators.
}
\description{
The helper function to generate the predictive probability of success
based on the difference in treatment group (\code{E}) and control or
standard of care (\code{SOC}) group.
}
\keyword{internal}
