% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprobDist.R
\name{h_predprobdist_single_arm}
\alias{h_predprobdist_single_arm}
\title{The predictive probability of success in single arm studies.}
\usage{
h_predprobdist_single_arm(
  x,
  mE,
  n,
  Nmax,
  delta,
  relativeDelta,
  parE,
  parS,
  weights,
  weightsS,
  thetaT
)
}
\arguments{
\item{x}{(\code{number}):\cr number of successes in the \code{E} group at interim.}

\item{mE}{(\code{number}):\cr number of successes in the remaining \code{Nmax-n} number of patients in the treatment \code{E} group.}

\item{Nmax}{(\code{number}):\cr maximum number of patients in the \code{E} group at final analysis.}

\item{delta}{(\code{number}):\cr difference between response rates to be met.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients. See \verb{[postprobDist()]}.}

\item{parE}{(\code{numeric}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{E} group. See details.}

\item{parS}{(\code{numeric}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution for the \code{S} group.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior.}

\item{weightsS}{(\code{numeric}):\cr weights for the \code{S} group.}

\item{thetaT}{(\code{number}):\cr threshold on the probability to be used.}

\item{density}{(\code{numeric}):\cr the beta binomial mixed density for future success in \code{Nmax-n} patients in the \code{E} group.}
}
\value{
A \code{list} is returned with names \code{result} for predictive probability and
\code{table} of numeric values with counts of responses in the remaining patients,
probabilities of these counts, corresponding probabilities to be above threshold,
and trial success indicators.
}
\description{
The helper function to generate the predictive probability of success based only on treatment group (\code{E})
as there is no control or standard of care (\code{S}), indicated by \code{NmaxControl == 0}.
}
\keyword{internal}
