% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logit}
\alias{logit}
\title{The Logit Function}
\usage{
logit(x)
}
\arguments{
\item{x}{(\code{numeric}):\cr probabilities to transform.}
}
\value{
The logit of \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This just calculates the logit.
}
\examples{
## If x = 0.2
## then logit(x) = log(0.2/(1-0.2)) = -1.386294
##
x <- 0.2
logit(x)

## Can also be a vector of probabilities
##
x <- seq(0.1, 0.9, 0.1)
logit(x)
}
