% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{myPlot}
\alias{myPlot}
\title{Plot the Beta distribution}
\usage{
myPlot(alpha, beta, ...)
}
\arguments{
\item{alpha}{(\code{number}):\cr first parameter of the Beta distribution}

\item{beta}{(\code{number}):\cr second parameter of the Beta distribution}
}
\value{
A beta distribution density plot
}
\description{
This function will plot the PDF of a beta distribution
}
\examples{
myPlot(alpha = 1, beta = 2)
myPlot(alpha = 1, beta = 1)
}
\keyword{graphics}
