% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{myPlotDiff}
\alias{myPlotDiff}
\title{Plot Diff Between two Beta distributions}
\usage{
myPlotDiff(parY, parX, cut_B = 0.2, cut_W = 0.1, shade = 1, note = 1, ...)
}
\arguments{
\item{parY}{non-negative parameters of the treatment Beta distribution.}

\item{parX}{non-negative parameters of the historical control Beta distribution}

\item{cut_B}{a meaningful improvement threshold}

\item{cut_W}{a poor improvement throshold}

\item{shade}{paint the two areas under the curve, default value=1 as "yes". other numbers stands for "no";}

\item{note}{show values of the colored area, default value=1 as "yes". other numbers stands for "no"}

\item{\dots}{additional arguments to \code{plot}}
}
\value{
nothing, only produces the plot as side effect
}
\description{
This function will plot the PDF of a diffience between two Beta distributions
}
\examples{
myPlotDiff(c(5, 10), c(2, 5), 0.2, 0.05, 1, 0)
}
\keyword{graphics}
