% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{ocPostprob}
\alias{ocPostprob}
\title{Calculate operating characteristics for posterior probability method}
\usage{
ocPostprob(
  nn,
  p,
  p0,
  p1,
  tL,
  tU,
  parE = c(1, 1),
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)
}
\arguments{
\item{nn}{vector of look locations for efficacy
(if futility looks should be different, please specify also \code{nnF})}

\item{p}{true rate (scenario)}

\item{p0}{lower efficacy threshold}

\item{p1}{upper efficacy threshold}

\item{tL}{probability threshold for being below p0}

\item{tU}{probability threshold for being above p1}

\item{parE}{beta parameters for the prior on the treatment proportion}

\item{ns}{number of simulations}

\item{nr}{generate random look locations? (not default)}

\item{d}{distance for random looks around the look locations in \code{nn}}

\item{nnF}{vector of look locations for futility
(default: same as efficacy)}
}
\value{
A list with the following elements:
oc: matrix with operating characteristics (see Details section)
Decision: vector of the decisions made in the simulated trials
(\code{TRUE} for success, \code{FALSE} for failure, \code{NA} for no
decision)
SampleSize: vector of the sample sizes in the simulated trials
nn: vector of look locations that was supplied
nnE: vector of efficacy look locations
nnF: vector of futility look locations
params: multiple parameters
}
\description{
The trial is stopped for efficacy if the posterior probability to be
above p1 is larger than tU, and stopped for futility if the posterior
probability to be below p0 is larger than tL.
}
\details{
Returned operating characteristics in a matrix include:
ExpectedN: expected number of patients in the trials
PrStopEarly: probability to stop the trial early (before reaching the
maximum sample size)
PrEarlyEff: probability to decide for efficacy early
PrEarlyFut: probability to decide for futility early
PrEfficacy: probability to decide for efficacy
PrFutility: probability to decide for futility
PrGrayZone: probability of no decision at the end ("gray zone")
}
\examples{
# operating characteristics for posterior probability method

# design details (example)
# multiple looks @ 10, 20, 30 patietns
# True response rate of the treatment group=40\%
# stop for futility: P(response rate < 20\% )> 60\%
# s top for efficacy: P(response rate > 30\% )> 80\%
# prior of treatment arm parE= Beta(1,1)
res1 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000
)
res1$oc

# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:
res2 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE
)
res2$oc

# now d is specified:
res3 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE, d = 5
)
res3$oc

# finally, we can also have separate specification of efficacy and
# futility analyses. E.g. here have futility decisions only at the final analysis:
res4 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE, d = 5,
  nnF = c(30)
)
res4$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.
}
