% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{ocPostprob}
\alias{ocPostprob}
\title{Operating Characteristics for Posterior Probability method}
\usage{
ocPostprob(
  nn,
  p,
  p0,
  p1,
  tL,
  tU,
  parE = c(1, 1),
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)
}
\arguments{
\item{nn}{(\code{numeric}):\cr sample size or sizes where study can be stopped for efficacy decision. If different for futility decision,
specify in \code{nnF}.}

\item{p}{(\code{number}):\cr assumed true rate of response.
true rate (scenario)}

\item{p0}{(``):\cr lower efficacy threshold.}

\item{p1}{(``):\cr upper efficacy threshold.}

\item{tL}{(``):\cr probability threshold for being below \code{p0}.}

\item{tU}{(``):\cr probability threshold for being above \code{p1}.}

\item{parE}{(\code{numeric}):\cr beta parameters for the prior on the treatment proportion.}

\item{ns}{(\code{number}):\cr number of simulations.}

\item{nr}{(\code{number}):\cr generate random look locations (not default)}

\item{d}{(\code{numeric}):\cr distance for random looks around the look locations in \code{nn}.}

\item{nnF}{(``):\cr sample size or sizes where study can be stopped for efficacy decision. If different for futility decision,
specify in \code{nnF}.}
}
\value{
A list with the following elements:
\itemize{
\item oc: matrix with operating characteristics (see Details section)
Decision: vector of the decisions made in the simulated trials
(\code{TRUE} for success, \code{FALSE} for failure, \code{NA} for no
decision)
SampleSize: vector of the sample sizes in the simulated trials
\item \code{nn}: vector of look locations that was supplied
\item \code{nnE}: vector of efficacy look locations
\item \code{nnF}: vector of futility look locations
\item \code{params}: multiple parameters
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate operating characteristics for posterior probability method.

The trial is stopped for efficacy if the posterior probability to be
above p1 is larger than tU, and stopped for futility if the posterior
probability to be below p0 is larger than tL:

Stop criteria for Efficacy : \code{P_E(p > p1) > tU}

Stop criteria for Futility \code{P_E(p < p0) < tL}

Resulting Operating Characteristics include the following:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability to decide for futility early
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: Probability of stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or grey decision zone
}
}
\examples{
# operating characteristics for posterior probability method

# design details (example)
# multiple looks @ 10, 20, 30 patients
# True response rate of the treatment group=40\%
# stop for futility: P(response rate < 20\% )> 60\%
# s top for efficacy: P(response rate > 30\% )> 80\%
# prior of treatment arm parE= Beta(1,1)
res1 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000
)
res1$oc

# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:
res2 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE
)
res2$oc

# now d is specified:
res3 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE, d = 5
)
res3$oc

# finally, we can also have separate specification of efficacy and
# futility analyses. E.g. here have futility decisions only at the final analysis:
res4 <- ocPostprob(
  nn = c(10, 20, 30), p = 0.4, p0 = 0.2, p1 = 0.3, tL = 0.6, tU = 0.8,
  parE = c(1, 1), ns = 1000, nr = TRUE, d = 5,
  nnF = c(30)
)
res4$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.
}
