% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprob.R
\name{ocPostprob}
\alias{ocPostprob}
\title{Operating Characteristics for Posterior Probability method}
\usage{
ocPostprob(
  nnE,
  truep,
  p0,
  p1,
  tL,
  tU,
  parE = c(1, 1),
  sim = 50000,
  wiggle = FALSE,
  nnF = nnE
)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate.}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{p1}{(\code{number}):\cr upper Efficacy threshold of response rate.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{wiggle}{(\code{flag}):\cr generate random look locations (not default).
if \code{TRUE}, optional to specify \code{dist} (see @details).}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}
}
\value{
A list with the following elements:
\itemize{
\item \code{oc}: matrix with operating characteristics (see @details section)
\item \code{nn}: vector of look locations that was supplied
\item \code{nnE}: vector of Efficacy look locations
\item \code{nnF}: vector of Futility look locations
\item \code{params}: user input arguments.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate operating characteristics for posterior probability method.

It is assumed that the true response rate is \code{truep}.
The trial is stopped for Efficacy if the posterior probability to be
above \code{p1} is larger than \code{tU}, and stopped for Futility if the posterior
probability to be below \code{p0} is larger than \code{tL}:

Stop criteria for Efficacy :

\code{Pr(truep > p1) > tU}

Stop criteria for Futility :

\code{Pr(truep < p0) > tL}

Resulting operating characteristics include the following:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\examples{
# For three looks of 10, 20 and 30 we have the following assumptions :
# True response rate or truep of the treatment group = 40\%
# The following are the Stop rules respectively :
# Look for Efficacy: Pr(truep > 30\% )> 80\%
# Look for Futility: Pr(truep < 20\% )> 60\%
# Prior of treatment arm parE= Beta(1,1).
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50, wiggle = TRUE, nnF = c(10, 20, 30)
)

res$oc

# We specify the distance in this example.
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50, wiggle = TRUE, nnF = c(10, 20, 30)
)

res$oc

# Here, nnE = nnF, and no wiggle room is allowed. Random distance also not supplied.
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50, wiggle = FALSE, nnF = c(10, 20, 30)
)

res$oc

# Here, we only have one Futility and Efficacy look or stop.
res <- ocPostprob(
  nnE = c(10), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50, wiggle = FALSE, nnF = 10
)

res$oc

# Here, we only have one Futility but many Efficacy looks or stop.
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50, wiggle = FALSE, nnF = 10
)

res$oc
}
