% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprobDist.R
\name{ocPostprobDist}
\alias{ocPostprobDist}
\title{Calculate Operating characteristics of Posterior Probability method
with Beta Prior on the Control or Standard of Care Arm}
\usage{
ocPostprobDist(
  nn,
  truep,
  deltaE,
  deltaF,
  relativeDelta = FALSE,
  tL,
  tU,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  wiggle = TRUE,
  sim = 50000,
  nnF = nn
)
}
\arguments{
\item{nn}{(\verb{number or numeric}):\cr the union of \code{nnE} and \code{nnF} (if futility analysis or looks exists) supplied.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{deltaE}{(\code{number}):\cr margin by which the response rate in the treatment group should be better
than in the standard of care of \code{group}. Delta for efficacy is used to
calculate \code{P(P_E > P_S + deltaE)} which should
exceed threshold \code{tU} to to stop for efficacy.
Note that this can also be negative, e.g. when non-inferiority is being assessed.
See note.}

\item{deltaF}{(\code{number}):\cr margin by which the response rate in the treatment group should be better
than in the standard of care of \code{group}. Delta for futility is used to
calculate  \code{P(P_E > P_S + deltaS)} which should
exceed threshold \code{tL} to stop for futility.
Note that this can also be negative, e.g. when non-inferiority is being assessed.
See note.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients
is included as the margin between treatment and control group. See note.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{S} group. See details.}

\item{wiggle}{(\code{flag}):\cr generate random look locations (not default).
if \code{TRUE}, optional to specify \code{dist} (see @details).}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}
}
\description{
The trial is stopped for efficacy if the posterior probability to be at
least \code{deltaE} better than the control is larger than \code{tU}, and stopped for
futility if the posterior probability to be at least \code{deltaF} worse than the
control is larger than \code{tL}. Otherwise the trial is continued, and at the
maximum sample size it may happen that no decision is made ("gray zone").
}
\details{
#' Stop criteria for Efficacy :

\code{P_e(p > p1 + deltaE) > tU}

Stop criteria for Futility :

\code{P_E(p < p0 + deltaF) > tL}

Resulting operating characteristics include the following:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\note{
\subsection{Delta :}{

The desired improvement is denoted as \code{delta}. There are two options in using \code{delta}.
The absolute case when \code{relativeDelta = FALSE} and relative as when \code{relativeDelta = TRUE}.
\enumerate{
\item The absolute case is when we define an absolute delta, greater than \code{P_S},
the response rate of the standard of care or control or \code{S} group such that
the posterior is \code{Pr(P_E > P_S + deltaE | data)} for efficacy looks
or \code{Pr(P_E > P_S + deltaF | data)} for futility looks.
\item In the relative case, we suppose that the treatment group's
response rate is assumed to be greater than \code{P_S + (1-P_S) * delta} such that
the posterior is \code{Pr(P_E > P_S + (1 - P_S) * deltaE | data)} for efficacy looks
or \code{Pr(P_E > P_S + (1 - P_S) * deltaF | data)} for futility looks.
}
}
}
\examples{
# operating characteristics for posterior probability method with beta prior on SOC

# design details
# multiple looks @ 10, 20, 30 patietns
# True response rate of the treatment group=0.4

# stop for efficacy (deltaE): P(pE > pS + deltaE) >tU
# stop for futility (deltaF): P(pE < pS - deltaF) >tL
# where pE is the response rate of treatment group, pS is the response rate of
# control group. Both of them are random variables.


res1 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100
)

res1$oc


# generate random look locations around 10,20,30 patients
# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:
res2 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100, nr = TRUE
)

res2$oc

# specify the distance for random looks around the look locations in nn (d=5 for illustration)
res3 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100,
  nr = TRUE,
  d = 5
)

res3$oc


# finally, we can also have separate specification of efficacy and futility analyses. E.g.
# here have futility decisions only at the final analysis:
res4 <- ocPostprobDist(
  nn = c(10, 20, 30), p = 0.4, deltaE = 0.1, deltaF = -0.1, tL = 0.6, tU = 0.6,
  parE = c(1, 1), parS = c(5, 25), ns = 100, nr = TRUE, d = 5, nnF = c(30)
)
res4$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.
}
