% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPostprobDist.R
\name{ocPostprobDist}
\alias{ocPostprobDist}
\title{Calculate Operating characteristics of Posterior Probability method
with Beta Prior on the Control or Standard of Care Arm}
\usage{
ocPostprobDist(
  nnE,
  truep,
  deltaE,
  deltaF,
  relativeDelta = FALSE,
  tL,
  tU,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  wiggle = TRUE,
  sim = 50000,
  nnF = nnE
)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision
specify in \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{deltaE}{(\code{number}):\cr margin of which the response rate in the treatment group should be better
than in the standard of care of \code{group}. Delta for Efficacy is used to
calculate \code{P(truep > P_S + deltaE)} which should
exceed threshold \code{tU} to to stop for Efficacy. \code{DeltaE} can also be negative in
non-inferiority setting. See also note.}

\item{deltaF}{(\code{number}):\cr margin of which the response rate in the treatment group should be better
than in the standard of care of \code{group}. Delta for Futility is used to
calculate  \code{P(truep > P_S - deltaF)} which should
exceed threshold \code{tL} to stop for Futility.\code{DeltaF} can also be negative in
non-inferiority setting. See also note.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients
is included as the margin between treatment and control group. See note.}

\item{tL}{(\code{number}):\cr posterior probability threshold for being below \code{p0}.}

\item{tU}{(\code{number}):\cr posterior probability threshold for being above \code{p1}.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{S} group. See details.}

\item{wiggle}{(\code{flag}):\cr generate random look locations (not default).
if \code{TRUE}, optional to specify \code{dist} (see @details).}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
The trial is stopped for Efficacy if the posterior probability of assumed
true response rate or \code{truep} is better than the control arm by at least
\code{deltaE}, is larger than \code{tU}. Analogously, The trial is stopped for Efficacy
if the posterior probability of assumed true response rate or \code{truep} is
better than the control arm by at least \code{deltaF}, is larger than \code{tL}
for Futility looks. If these criteria are not met, the trial is continued
and at the maximum sample size, it may reach no decision or "gray zone".The
following represents the stopping criteria for Efficacy and Futility
respectively.

Stop criteria for Efficacy :

\code{Pr(truep > P_S + deltaE) > tU}

Stop criteria for Futility :

\code{Pr(truep < P_S - deltaF) > tL}

Where \code{truep} is the assumed true rate of response and \code{p1} and \code{p0} respectively are
the thresholds for Efficacy and Futility respectively.

Resulting operating characteristics include the following:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\note{
\subsection{Delta :}{

The desired improvement is denoted as \code{deltaE} or \code{deltaF} and two options
exist in its use: The absolute case when \code{relativeDelta = FALSE} and
relative as when \code{relativeDelta = TRUE}.
\enumerate{
\item The absolute case is when we define an absolute delta, greater than \code{P_S},
the response rate of the standard of care or control or \code{S} group such that
the posterior is \code{Pr(truep > P_S + deltaE | data)} for Efficacy looks
or \code{Pr(truep > P_S + deltaF | data)} for Futility looks.
\item In the relative case, we suppose that the treatment group's
response rate is assumed to be greater than \code{P_S + (1-P_S) * delta} such that
the posterior is \code{Pr(truep > P_S + (1 - P_S) * deltaE | data)} for Efficacy looks
or \code{Pr(truep > P_S + (1 - P_S) * deltaF | data)} for Futility looks.
}
}
}
\examples{
# operating characteristics for posterior probability method with beta prior on SOC
# For three looks of 10, 20 and 30 we have the following assumptions :
# True response rate or truep of the treatment group = 40\%
# The following are the Go and Stop rules respectively :
# Prior of treatment arm parE= Beta(1,1).
# stop for efficacy (deltaE): Pr(truep > P_S + deltaE) > tU
# stop for futility (deltaF): Pr(truep < P_S - deltaF) > tL
# Without random distance allowed for Futility and Efficacy Looks.
res1 <- ocPostprobDist(
  nnE = c(10, 20, 30),
  truep = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  sim = 5000,
  wiggle = FALSE
)

res1$oc

# Allow random distance for Efficacy and Futility looks.
res2 <- ocPostprobDist(
  nnE = c(10, 20, 30),
  truep = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  sim = 100,
  wiggle = TRUE
)

res2$oc

# Allow Futility analysis only at the end.
res4 <- ocPostprobDist(
  nnE = c(10, 20, 30),
  truep = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  sim = 100,
  wiggle = TRUE,
  nnF = 30
)

res4$oc
}
