% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{ocPredprob}
\alias{ocPredprob}
\title{Operating Characteristics for Predictive Probability method}
\usage{
ocPredprob(
  nnE,
  truep,
  p0,
  p1 = p0,
  tT = 1 - tF,
  tF = 1 - tT,
  phiL = 1 - phiFu,
  phiU,
  phiFu = 1 - phiL,
  parE = c(1, 1),
  sim = 50000,
  wiggle = FALSE,
  nnF = nnE,
  decision1 = TRUE
)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{p0}{(\code{number}):\cr lower Futility threshold of response rate.}

\item{p1}{(\code{number}):\cr upper Futility threshold of response rate.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{tF}{(\code{number}):\cr threshold of which assumed \code{truep} does not exceed threshold of \code{p1}.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{phiFu}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{wiggle}{(\code{flag}):\cr generate random look locations (not default).
if \code{TRUE}, optional to specify \code{dist} (see @details).}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision.}

\item{decision1}{(\code{flag}):\cr Flag if \code{decision1 = TRUE} then Decision 1 will be used.}
}
\value{
A list with the following elements:
\itemize{
\item \code{oc}: matrix with operating characteristics (see @details section)
\item \code{nn}: vector of look locations that was supplied
\item \code{nnE}: vector of Efficacy look locations
\item \code{nnF}: vector of Futility look locations
\item \code{params}: multiple parameters
\item \code{decision1} : flag \code{TRUE} to indicate that Decision 1 was used
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate operating characteristics for predictive probability method.

It is assumed that the true response rate is \code{truep}. Trials can stop for Futility
or Efficacy. Trials can also stop at Interim or Final for Futility or Efficacy.

There are two variations of decision rule, Decision 1 and Decision 2, to
evaluate decision at Interim or Final, for Futility or Efficacy. Decision 1 is
used when \code{decision1 == TRUE} which is the default setting.

The criteria for Decision 1 for Interim looks are :
\itemize{
\item interim GO =  P(successful trial at final) > phiU
\item interim STOP = P(successful trial at final) < phiL
}

The criteria for Decision 1 for Final looks are:
\itemize{
\item Final GO = P(truep > p0 | data) => tT
\item Final STOP = P(truep > p0 | data ) < tT
}

The criteria for Decision 2 for Interim looks are :
\itemize{
\item Interim GO : P ( success at final) > phiU
\item Interim STOP : P (failure at final ) > phiFu
}

The criteria for Decision 2 for Futility looks are :
\itemize{
\item Final GO = P(truep > p0) > tT
\item Final STOP = P(truep < p1) > tF
}

Resulting operating characteristics include the following:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
}
\details{
\code{ExpectedN} is an average of the simulated sample sizes.
If \code{wiggle = TRUE}, one can specify \code{dist}, though the algorithm will generate it if \code{dist = NULL}.
If \code{nnF = NULL}, no Futility or decision to Stop will be analysed. Note that \code{nnF = c(0)} is equivalent.
As default, \code{nnF} is set to the identical looks of \code{nnE}, and if \code{wiggle = TRUE}, all looks are the same, e.g.
\code{nnE = nnF} when wiggle and distance is applied.
}
\examples{
# For various Look locations at Interim and Final, we have the following assumptions :
# True response rate or truep of the treatment group = 40\%
# The following are the Final Stop rules respectively :
# - Final look for Efficacy: Pr( response rate > 25\% ) > 60\%
# - Final look for Futility: Pr( response rate < 25\% ) < 60\%
# - Interim look for Efficacy: Pr( success at final ) > 80\%
# - Interim look for Futility: Pr( failure at final ) < 20\%, or > 80 \% for Decision 2
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.

set.seed(20)
result <- ocPredprob(
  nnE = c(10, 20),
  truep = 0.4,
  p0 = 0.25,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 100,
  wiggle = FALSE,
  decision1 = TRUE
)
result$oc

# For when random distance is allowed, i.e. `wiggle = TRUE` :
ocPredprob(
  nnE = c(10, 20),
  truep = 0.6,
  p0 = 0.2, # no p1 as this is decision 1
  tT = 0.6, # no tF as it's for decision 2
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = TRUE,
  nnF = c(10, 20),
  decision1 = TRUE
)

# For Decision 2, where the Futility stop rules in interim and final are
# P(failure at final) > PhiFu
# P(response rate > p1) > tF
ocPredprob(
  nnE = c(10, 20),
  truep = 0.6,
  p0 = 0.2,
  p1 = 0.2,
  tT = 0.6,
  tF = 0.4,
  phiU = 0.2,
  phiFu = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = FALSE,
  nnF = c(10, 20),
  decision1 = FALSE
)

# Separate Futility and Efficacy looks at interim and final can be performed:
ocPredprob(
  nnE = c(10, 25, 30),
  truep = 0.6,
  p0 = 0.2,
  p1 = 0.2,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = FALSE,
  nnF = c(10, 15, 20),
  decision1 = TRUE
)

# Separate Futility and Efficacy looks at interim and final can be performed:
ocPredprob(
  nnE = c(10, 25, 30),
  truep = 0.6,
  p0 = 0.2,
  p1 = 0.2,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(1, 1),
  sim = 50,
  wiggle = TRUE,
  nnF = c(10, 15, 20),
  decision1 = TRUE
)

# Futility only at final with non-uniform beta prior parE:
ocPredprob(
  nnE = c(10, 25, 30),
  truep = 0.6,
  p0 = 0.2,
  p1 = 0.2,
  tT = 0.6,
  phiL = 0.2,
  phiU = 0.8,
  parE = c(11, 19),
  sim = 50,
  wiggle = TRUE,
  nnF = 30,
  decision1 = TRUE
)
}
