% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprob.R
\name{ocPredprob}
\alias{ocPredprob}
\title{Calculate operating characteristics for predictive probability method
(gray zone allowed in the final analysis)}
\usage{
ocPredprob(
  nn,
  p,
  p0,
  p1 = p0,
  tT,
  tFu = 1 - tT,
  phiL = 1 - phiFu,
  phiU,
  phiFu = 1 - phiL,
  parE = c(1, 1),
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)
}
\arguments{
\item{nn}{vector of look locations for efficacy
(if futility looks should be different, please specify also \code{nnF})}

\item{p}{true rate (scenario)}

\item{p0}{threshold on the response rate}

\item{p1}{futility threshold on the response rate. Specify only when decision rule 2 is used.}

\item{tT}{threshold for the probability to be above the response rate p0
at the end of the trial}

\item{tFu}{threshold for the probability to be under the response rate p1
at the end of the trial. Specify only when decision rule 2 is used.}

\item{phiL}{lower threshold on the predictive probability}

\item{phiU}{upper threshold on the predictive probability}

\item{phiFu}{threshold on the predictive probability for futility. Specify only
when decision rule 2 is used. phiL argument should be skipped in this case.}

\item{parE}{beta parameters for the prior on the treatment proportion}

\item{ns}{number of simulations}

\item{nr}{generate random look locations? (not default)}

\item{d}{distance for random looks around the look locations in \code{nn}}

\item{nnF}{vector of look locations for futility
(default: same as efficacy)}
}
\value{
A list with the following elements:
oc: matrix with operating characteristics (see Details section)
Decision: vector of the decisions made in the simulated trials
(\code{TRUE} for success, \code{FALSE} for failure)
SampleSize: vector of the sample sizes in the simulated trials
nn: vector of look locations
nnE: vector of efficacy look locations
nnF: vector of futility look locations
params: multiple parameters
}
\description{
Decision rule 1:The trial is stopped for efficacy if the predictive probability of a
successful trial is larger than phiU, and stopped for futility if it is below
phiL. A trial is successful if after the maximum number of patients the
posterior probability of the treatment having more than p0 response rate is
above tT. Otherwise the decision is "failure". In this case, there is no gray zone.
}
\details{
Decision rule 2:A variation can be requested when skipping the argument phiL and utilizing the arguments
p1, tFu & PhiFu. The trial can be stopped for futility if the predictive
probability of an unsuccessful trial is larger than phiFu. In this case, the
decision is "failure" when the posterior probability of having the treatment
response rate at most p1 is above tFu at the maximum number of patients.

Returned operating characteristics in a matrix include:
ExpectedN: expected number of patients in the trials
PrStopEarly: probability to stop the trial early (before reaching the
maximum sample size)
PrEarlyEff: probability to decide for efficacy early
PrEarlyFut: probability to decide for futility early
PrEfficacy: probability to decide for efficacy
PrFutility: probability to decide for futility
PrGrayZone: probability of no decision at the end ("gray zone")
}
\examples{
## to examine operating characteristics of a phase Ib design with 2 interim analyses
## (sample size @ 10, 20 using predictive probability) and one final analysis (sample size @ 30);

# design details
# multiple looks @ 10, 20, 30 patietns
# True response rate of the treatment group=40\%
# control response rate=25\%

# denote treatment response rate as RRS
# trial success is defined as: P(RRS>control)>tT
set.seed(20)
res1 <- ocPredprob(c(10, 20, 30),
  p = 0.4, p0 = 0.25, tT = 0.6, phiL = 0.2,
  phiU = 0.8, parE = c(1, 1), ns = 100
)
res1$oc


# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:

res2 <- ocPredprob(c(10, 20, 30), 0.4,
  p0 = 0.25, tT = 0.6, phiL = 0.2, phiU = 0.8,
  parE = c(1, 1), ns = 100, nr = TRUE
)
res2$oc

# now d is specified. note that d has to be a positive value;
res3 <- ocPredprob(c(10, 20, 30), 0.4,
  p0 = 0.25, tT = 0.6, phiL = 0.2, phiU = 0.8,
  parE = c(1, 1), ns = 100, nr = TRUE, d = 5
)
res3$oc


# finally, we can also have separate specification of efficacy and futility analyses. E.g.
# here have futility decisions only at the final analysis:


res4 <- ocPredprob(c(10, 20, 30), 0.4,
  p0 = 0.25, tT = 0.6, phiL = 0.2, phiU = 0.8,
  parE = c(1, 1), ns = 100, nr = TRUE, d = 5, nnF = c(30)
)
res4$oc
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.

# 2)
# to examine operating characteristics of a phase Ib design with a seperate
# rule for futility at final analysis
# denote treatment response rate as RRS
# trial success is defined as: P(RRS>p0)>tT
# trial failure is defined as: P(RRS<p1)>tFu
# (gray zone could occur in the final analysis)

# when tFu=1-tT and phiFu=1-phiL, this decision rule (res5) is the same as the res1

set.seed(20)
res5 <- ocPredprob(c(10, 20, 30),
  p = 0.4, p0 = 0.25, tT = 0.6, tFu = 0.4, phiFu = 0.8,
  phiU = 0.8, parE = c(1, 1), ns = 100
)
res5$oc
}
