% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocPredprobDist.R
\name{ocPredprobDist}
\alias{ocPredprobDist}
\title{Calculate operating characteristics for predictive probability method
with beta prior on SOC response rate (gray zone allowed in the final analysis for Decision 2)}
\usage{
ocPredprobDist(
  nnE,
  truep,
  deltaE,
  deltaF,
  phiU,
  xS = 0,
  nS = 0,
  relativeDelta = FALSE,
  tT = 1 - tF,
  tF = 1 - tT,
  phiL = 1 - phiFu,
  phiFu = 1 - phiL,
  parE = c(1, 1),
  parS = c(1, 1),
  weights = weights,
  weightsS = weightsS,
  sim = 50000,
  wiggle = FALSE,
  nnF = nnE,
  decision1 = TRUE
)
}
\arguments{
\item{nnE}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision,
specify in \code{nnF}.}

\item{truep}{(\code{number}):\cr assumed true response rate or true rate (scenario).}

\item{deltaE}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Efficacy looks only.
Note that this can also be negative as well.}

\item{deltaF}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the standard of care or control or \code{S} group in Futility looks only.
Note that this can also be negative as well.}

\item{phiU}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{xS}{(\code{number}):\cr number of successes in the \code{S} group at interim.}

\item{nS}{(\code{number}):\cr number of patients in the \code{E} group at interim.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the \code{S} non-responding patients. See \code{\link[=postprobDist]{postprobDist()}}.}

\item{tT}{(\code{number}):\cr threshold of which assumed \code{truep} exceeds acceptable threshold of \code{p0}.}

\item{tF}{(\code{number}):\cr threshold of which assumed \code{truep} does not exceed threshold of \code{p1}.}

\item{phiL}{(\code{number}):\cr lower threshold on the predictive probability.}

\item{phiFu}{(\code{number}):\cr upper threshold on the predictive probability.}

\item{parE}{(\code{numeric}):\cr alpha and beta parameters for the prior on the treatment population.
Default set at alpha = 1, beta = 1, or uniform prior.}

\item{parS}{(\code{numeric}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution for the \code{S} group.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior.}

\item{weightsS}{(\code{numeric}):\cr weights for the \code{S} group.}

\item{sim}{(\code{number}):\cr number of simulations.}

\item{wiggle}{(\code{flag}):\cr generate random look locations (not default). If \code{wiggle = TRUE} and \code{nnE = nnF}, then all wiggled
looks are the same between \code{nnE} and \code{nnF}.}

\item{nnF}{(\code{numeric}):\cr sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.}

\item{decision1}{(\code{flag}):\cr Flag if \code{decision1 = TRUE} then Decision 1 rules will be used, otherwise Decision 2 rules will be used.}
}
\value{
A list with the following elements:
\itemize{
\item \code{oc}: matrix with operating characteristics with the following details:
\itemize{
\item \code{ExpectedN}: expected number of patients in the trials
\item \code{PrStopEarly}: probability to stop the trial early (before reaching the
maximum sample size)
\item \code{PrEarlyEff}: probability of Early Go decision
\item \code{PrEarlyFut}: probability of for Early Stop decision
\item \code{PrEfficacy}: probability of Go decision
\item \code{PrFutility}: probability of Stop decision
\item \code{PrGrayZone}: probability between Go and Stop ,"Evaluate" or Gray decision zone
}
\item \code{Decision} : numeric of results with \code{TRUE} as Go decision, \code{FALSE} as Stop and \code{NA} as gray zone.
\item \code{SampleSize} : numeric of sample sizes from \code{nnE} or \code{nnF} or both.
\item \code{wiggled_nnE} : user input for \code{nnE} with random distance applied.
\item \code{wiggled_nnF} : user input for \code{nnF} with random distance applied.
\item \code{wiggled_dist} : magnitude of random distance applied in order of input looks.
\item \code{params} : all user input arguments.
}
}
\description{
For formulas and expressions for predictive probability, refer to \code{\link[=predprobDist]{predprobDist()}}.
}
\details{
The rules for Stop, Go and Gray Zone (where applicable) are the same as in \code{\link[=ocPredprob]{ocPredprob()}}.
However when compared to \code{\link[=ocPredprob]{ocPredprob()}}, where the predictive probability of \verb{response rate > p0}
was used to reach decisions, here the \verb{response rate - p0 > 0} is replaced by \verb{response rate - p0 > delta},
and the reason is that \code{p0} is now also uncertain, which is quantified by a posterior distribution.
To therefore understand the margin of difference between the experimental and control group, refer
to the two ways of calculating delta in \code{\link[=postprobDist]{postprobDist()}}.

The margin \code{delta} from \code{\link[=postprobDist]{postprobDist()}} is no longer used however where a Go decision is evaluated,
the margin \code{deltaE} is instead employed, and where a Stop decision is evaluated,
the margin \code{deltaF} is instead employed in lieu of \code{delta}.
Decision 1 and Decision 2 share the same Go rule for interim and final,
with the margin of difference is accounted for in the following manner for the final looks:
\itemize{
\item \code{Pr(P_E > p0 + deltaE | data) > tT } for the absolute case
\item \code{Pr(P_E > (1-p0)*deltaE | data) > tT} for the relative case
}

As mentioned, the rules of \code{\link[=ocPredprobDist]{ocPredprobDist()}} and \code{\link[=ocPredprob]{ocPredprob()}} for both Decision 1 and Decision 2 are the same,
and the margin of difference differ depending on whether it is a Go or Stop decision
(\code{deltaE} and \code{deltaF} is employed instead of \code{delta}).
We highlight here another distinction between \code{\link[=ocPredprob]{ocPredprob()}} and \code{\link[=ocPredprobDist]{ocPredprobDist()}},
seen in the evaluation for the final futility look in \code{\link[=ocPredprobDist]{ocPredprobDist()}} where \code{p1} is used instead of \code{p0} :
\itemize{
\item \code{Pr(P_E < p1 + deltaF | data) > tF} for the absolute case
\item \code{Pr(P_E < (1-p1)*deltaF | data) > tF} for the relative case
}
}
\examples{
# Here we illustrate an example for Decision 1 with the following assumptions :
# Efficacy Looks and Futility looks are identical at sample size of 10, 20 and 30.
# True response rate or truep of the treatment group = 40\%
# Desired difference to Standard of Care for Efficacy and Futility = 10\%
# Delta calculation is absolute case. The following are the Final Stop rules respectively :
# - Final look for Efficacy: Pr( response rate + deltaE > 25\% ) > 60\% or P(response rate + deltaE > p0) > tT
# - Final look for Futility: Pr( response rate + deltaF < 25\% ) < 60\% or P(response rate + deltaF > p0) < tT
# - Interim look for Efficacy: Pr( success at final ) > 80\% or P(success at final) > phiU
# - Interim look for Futility: Pr( failure at final ) < 20\% or P(success at final) < phiL
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.

set.seed(20)
result <- ocPredprobDist(
  nnE = c(10, 20, 30),
  truep = 0.40,
  deltaE = 0.10,
  deltaF = 0.10,
  relativeDelta = FALSE,
  tT = 0.6,
  phiU = 0.80,
  phiL = 0.20,
  parE = c(1, 1),
  parS = c(5, 25),
  weights = 1,
  weightsS = 1,
  sim = 50,
  wiggle = FALSE,
  decision1 = TRUE
)
result$oc


# Here we illustrate an example for Decision 1 with the following assumptions :
# Efficacy Looks are at sample size of 10, 20 and 30 where sample looks are allowed to wiggle.
# Futility is only at sample size of 20 and 30. We won't do a futility at 10 patients.
# True response rate or truep of the treatment group = 40\%
# Desired difference to Standard of Care for Efficacy and Futility is 10\% and -10\% respectively.
# Grey zone occurs due to different posterior probability distribution in the Efficacy and Futility rules.
# Delta calculation is absolute case. The following are the Final Stop rules respectively :
# - Final look for Efficacy: Pr( response rate + deltaE > 25\% ) > 60\% or P(response rate + deltaE > p0) > tT
# - Final look for Futility: Pr( response rate + deltaF < 25\% ) < 60\% or P(response rate + deltaF > p0) < tT
# - Interim look for Efficacy: Pr( success at final ) > 80\% or P(success at final) > phiU
# - Interim look for Futility: Pr( failure at final ) < 20\% or P(success at final) < phiL
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.
#
set.seed(20)
result <- ocPredprobDist(
  nnE = c(10, 20, 30),
  truep = 0.40,
  deltaE = 0.10,
  deltaF = -0.10,
  relativeDelta = FALSE,
  tT = 0.6,
  phiU = 0.80,
  phiL = 0.20,
  parE = c(1, 1),
  parS = c(5, 25),
  weights = 1,
  weightsS = 1,
  sim = 50,
  nnF = c(20, 30),
  wiggle = TRUE,
  decision1 = TRUE
)
result$oc

# Here we illustrate an example for Decision 2 with the following assumptions :
# Efficacy Looks and Futility looks are identical at sample size of 10, 20 and 30
# where sample looks are allowed to wiggle.
# True response rate or truep of the treatment group = 40\%
# Desired difference to Standard of Care for Efficacy and Futility = 50\%
# Delta calculation is absolute case. The following are the Final Stop rules respectively :
# - Final look for Efficacy: Pr( response rate + deltaE > 25\% ) > 60\% or P(response rate + deltaE > p0) > tT
# - Final look for Futility: Pr( response rate + deltaF < 25\% ) < 60\% or P(response rate + deltaF > p0) < tT
# - Interim look for Efficacy: Pr( success at final ) > 80\% or P(success at final) > phiU
# - Interim look for Futility: Pr( failure at final ) < 20\% or P(success at final) < phiL
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.

set.seed(20)
result <- ocPredprobDist(
  nnE = c(10, 20, 30),
  truep = 0.40,
  deltaE = 0.5,
  deltaF = 0.5,
  relativeDelta = FALSE,
  tT = 0.6,
  phiU = 0.80,
  phiFu = 0.7,
  parE = c(1, 1),
  parS = c(5, 25),
  weights = 1,
  weightsS = 1,
  sim = 50,
  nnF = c(10, 20, 30),
  wiggle = TRUE,
  decision1 = FALSE
)
result$oc

# Here we illustrate an example for Decision 2 with the following assumptions :
# Efficacy Looks and Futility looks are identical at sample size of 10, 20 and 30
# where wiggle is allowed.
# True response rate or truep of the treatment group = 40\%
# Desired difference to Standard of Care for Efficacy and Futility = 50\%
# Delta calculation is relative case. The following are the Final Stop rules respectively :
# - Final look for Efficacy: P( P_S + (1-P_S)*deltaE > 25\% ) > 60\% or P( P_S + (1-P_S)*deltaE > p0) > tT
# - Final look for Futility: P( P_S + (1-P_S)*deltaEF < 25\% ) < 60\% or P( P_S + (1-P_S)*deltaF > p0) < tT
# - Interim look for Efficacy: P( success at final ) > 80\% or P(success at final) > phiU
# - Interim look for Futility: P( failure at final ) < 20\% or P(success at final) < phiL
# We assume a prior of treatment arm parE = Beta(1,1), unless otherwise indicated.

set.seed(20)
result <- ocPredprobDist(
  nnE = c(10, 20, 30),
  truep = 0.40,
  deltaE = 0.5,
  deltaF = 0.5,
  relativeDelta = TRUE,
  tT = 0.6,
  phiU = 0.80,
  phiFu = 0.7,
  parE = c(1, 1),
  parS = c(5, 25),
  weights = 1,
  weightsS = 1,
  sim = 50,
  nnF = c(10, 20, 30),
  wiggle = TRUE,
  decision1 = FALSE
)
result$oc
}
