% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{pbetaMix}
\alias{pbetaMix}
\title{Beta-Mixture CDF}
\usage{
pbetaMix(q, par, weights, lower.tail = TRUE)
}
\arguments{
\item{q}{(\code{number}):\cr the abscissa.}

\item{par}{(\code{matrix}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior which add up to 1.}

\item{lower.tail}{(\code{flag}):\cr if \code{TRUE} (default), probabilities are \code{P[X <= x]},
and otherwise \code{P[X > x]}.}
}
\value{
The (one minus) cdf value
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the CDF of the Beta-Mixture distribution.
}
\note{
\code{q} can be a vector.
}
\examples{
pbetaMix(q = 0.3, par = rbind(c(0.2, 0.4)), weights = 1)

# Can get the one minus CDF values.
pbetaMix(q = 0.3, par = rbind(c(0.2, 0.4)), weights = 1, lower.tail = FALSE)

# With 2 mixture components
pbetaMix(
  q = 0.3, par = rbind(c(0.2, 0.4), c(1, 1)),
  weights = c(0.6, 0.4)
)

# Can also specify x as a vector.
pbetaMix(
  q = seq(0, 1, .01), par = rbind(c(0.2, 0.4), c(1, 1)),
  weights = c(0.6, 0.4)
)
}
