% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase1b-package.R
\docType{package}
\name{Phase1b-package}
\alias{phase1b-package}
\alias{Phase1b-package}
\alias{Phase1b}
\alias{phase1b}
\title{Calculations for decisions on Phase 1b clinical trials}
\description{
Phase 1b tools
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/genentech/phase1b/}
  \item Report bugs at \url{https://github.com/genentech/phase1b/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanes Bove \email{daniel@rconis.com}

Authors:
\itemize{
  \item Markus Elze \email{markus.elze@roche.com}
  \item Tony Pourmohamad \email{pourmoht@gene.com}
  \item Jiawen Zhu \email{zhu.jiawen@gene.com}
  \item James Lymp
  \item Anastasia Teterina \email{anastasia.teterina@roche.com}
  \item Audrey Yeo \email{audrey.yeo@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche Ltd [copyright holder, funder]
}

}
\keyword{package}
