% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBeta}
\alias{plotBeta}
\title{Plot the Beta distribution}
\usage{
plotBeta(alpha, beta)
}
\arguments{
\item{alpha}{(\code{number}):\cr first parameter of the Beta distribution}

\item{beta}{(\code{number}):\cr second parameter of the Beta distribution}
}
\value{
A beta distribution density plot
}
\description{
This function will plot a PDF of a beta distribution
}
\examples{
# plotBeta
plotBeta(alpha = 4, beta = 5)
plotBeta(alpha = 1, beta = 1)
}
\keyword{graphics}
