% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBetaDiff}
\alias{plotBetaDiff}
\title{Plot Diff Between two Beta distributions}
\usage{
plotBetaDiff(
  parY,
  parX,
  Go_cut = 0.2,
  Stop_cut = 0.1,
  shade = TRUE,
  note = TRUE
)
}
\arguments{
\item{parY}{(\code{numeric}):\cr non-negative parameters of the treatment Beta distribution.}

\item{parX}{(\code{numeric}):\cr non-negative parameters of the historical control Beta distribution}

\item{shade}{(\code{flag}):\cr paint the two areas under the curve, default value = TRUE}

\item{note}{(\code{flag}):\cr show values of the colored area, default value = TRUE}

\item{cut_B}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{cut_W}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{...}{(``):\cr additional arguments to \code{ggplot()}}
}
\value{
a ggplot object
}
\description{
This function will plot the PDF of a difference between two Beta distributions
}
\examples{
myPlotDiff(
  parY = c(5, 10),
  parX = c(2, 5),
  cut_B = 0.2, # a meaningful improvement threshold
  cut_W = 0.05, # a poor improvement threshold
  shade = 1, # paint the two areas under the curve, default: yes. other numbers stands for "no";
  note = 0
) # show values of the colored area, default: yes. other numbers stands for "no";
}
\keyword{graphics}
