% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBetaDiff}
\alias{plotBetaDiff}
\title{Plot difference Between two Beta distributions}
\usage{
plotBetaDiff(
  parX,
  parY,
  go_cut = 0.2,
  stop_cut = 0.1,
  shade = TRUE,
  note = TRUE
)
}
\arguments{
\item{parX}{(\code{numeric}):\cr non-negative parameters of the control Beta distribution}

\item{parY}{(\code{numeric}):\cr non-negative parameters of the treatment Beta distribution.}

\item{go_cut}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{stop_cut}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{shade}{(\code{flag}):\cr paint the two areas under the curve, default value = TRUE}

\item{note}{(\code{flag}):\cr show values of the colored area, default value = TRUE}
}
\value{
a ggplot object
}
\description{
This function will plot the PDF of a difference between two Beta distributions
}
\examples{
# The beta distribution and acceptable bounds for
# a meaningful improvement of 0.30 and worsening of 0.1
parX <- c(1, 52) # prior  parameters of control or SOC
parY <- c(5.5, 20.5) # prior  parameters of experimental arm
plotBetaDiff(
  parX = parX,
  parY = parY,
  go_cut = 0.3,
  stop_cut = 0.1, # below a difference of 10\%, is an unsuccessful trial
  shade = TRUE,
  note = TRUE
)

# a larger Go_cut with uniform prior
plotBetaDiff(
  parX = c(1, 1), # prior  parameters for experimental arm
  parY = c(1, 1), # prior parameters for control or SOC arm
  go_cut = 0.3,
  stop_cut = 0.1, # below a difference of 10\%, is an unsuccessful trial
  shade = TRUE,
  note = TRUE
)
# first example
sumBetadiff(
  parX = parX,
  parY = parY,
  coverage = 0.9,
  go_cut = 0.3,
  stop_cut = 0.1
)
# second example
sumBetadiff(
  parX = c(1, 1),
  parY = c(1, 1),
  coverage = 0.9,
  go_cut = 0.3,
  stop_cut = 0.1
)
}
\keyword{graphics}
