% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBetaDiff}
\alias{plotBetaDiff}
\title{Plot difference Between two Beta distributions}
\usage{
plotBetaDiff(
  parX,
  parY,
  go_cut = 0.2,
  stop_cut = 0.1,
  shade = TRUE,
  note = TRUE
)
}
\arguments{
\item{parX}{(\code{numeric}):\cr non-negative parameters of the control Beta distribution}

\item{parY}{(\code{numeric}):\cr non-negative parameters of the treatment Beta distribution.}

\item{go_cut}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{stop_cut}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{shade}{(\code{flag}):\cr paint the two areas under the curve, default value = TRUE}

\item{note}{(\code{flag}):\cr show values of the colored area, default value = TRUE}
}
\value{
a ggplot object
}
\description{
This function will plot the PDF of a difference between two Beta distributions
}
\examples{
# The following examples use these parameters:
parX <- c(1, 52)
parY <- c(5.5, 20.5)

# Calculate probability based on quantile `Q(Z) =< 0.122838`
pbetadiff(
  q = 0.122838,
  parY = parY,
  parX = parX
)

# Calculate probability based on quantile `Q(Z) =< 0.5`
pbetadiff(
  q = 0.5,
  parY = parY,
  parX = parX
)
}
\keyword{graphics}
