% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecision.R
\name{plotDecision}
\alias{plotDecision}
\title{Plot a summary plot corresponding to the sumTable output}
\usage{
plotDecision(data, Pos_cut, Neg_cut)
}
\arguments{
\item{data}{the output object of \code{\link{sumTable}}}

\item{Pos_cut}{a cut off for the prob of a meaningful improvement}

\item{Neg_cut}{a cut off for the prob of a poor improvement}
}
\value{
the \code{data} item which was imputed to the function
}
\description{
This function will return a plot showing a curve of the prob of a meaningful improvement over estunated diff
and a curve of the prob of a poor improvement over estunated diff
}
\examples{
# sumTable ----
sumTable(
  x = 10,
  n = 20,
  parX = c(1, 1),
  go_cut = 0.8,
  stop_cut = 0.4
)

# plotting more results
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    n = 25,
    parX = c(1, 52),
    go_cut = 0.6,
    stop_cut = 0.2
  )
)

# plotDecision(summaries, Pos_cut = 60, Neg_cut = 60)
}
\keyword{graphics}
