% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecision.R
\name{plotDecision}
\alias{plotDecision}
\title{Plot a summary plot corresponding to the sumTable output}
\usage{
plotDecision(data, Pos_cut, Neg_cut)
}
\arguments{
\item{data}{the output object of \code{\link{sumTable}}}

\item{Pos_cut}{a cut off for the prob of a meaningful improvement}

\item{Neg_cut}{a cut off for the prob of a poor improvement}
}
\value{
the \code{data} item which was imputed to the function
}
\description{
This function will return a plot showing a curve of the prob of a meaningful improvement over estunated diff
and a curve of the prob of a poor improvement over estunated diff
}
\examples{
summaries <- do.call(cbind, lapply(c(0:8), sumTable, TotalSample = 25, parX = c(1, 52), cut_B = 0.2, cut_W = 0.05))

plotDecision(summaries, Pos_cut = 60, Neg_cut = 60)
}
\keyword{graphics}
