% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecision.R
\name{plotDecision}
\alias{plotDecision}
\title{Plot a summary plot corresponding to the sumTable output}
\usage{
plotDecision(data, efficacious_prob, futile_prob)
}
\arguments{
\item{data}{(\code{data.frame}):\cr sourced \code{\link{data.frame}} from \link{(sumTable)}}

\item{efficacious_prob}{(\code{number}):\cr a cut off for the probability of a meaningful improvement}

\item{futile_prob}{(\code{number}):\cr a cut off for the probability of a poor improvement}
}
\value{
the \code{\link{data}} item which was imputed to the function
}
\description{
This function will return a plot showing a curve of the prob of a meaningful improvement over estimated diff
and a curve of the prob of a poor improvement over estimated diff
}
\examples{
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    n = 25,
    parX = c(1, 52),
    go_cut = 0.2,
    stop_cut = 0.05
  )
)

plotDecision(summaries, efficacious_prob = 60, futile_prob = 60)

# plotting different criteria
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    n = 25,
    parX = c(1, 52),
    # density when P( diff > 20\% | B(1, 52) for control and B(0.5, 0.5) for treatment) :
    go_cut = 0.2,
    # density when P( diff < 10\% | B(1, 52) for control and B(0.5, 0.5) for treatment) :
    stop_cut = 0.1
  )
)

plotDecision(summaries, efficacious_prob = 60, futile_prob = 80)
}
\keyword{graphics}
