% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprob}
\alias{postprob}
\title{Posterior Probability of Efficacy Given Beta-Mixture Prior}
\usage{
postprob(x, n, p, parE = c(1, 1), weights, betamixPost, log.p = FALSE)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes.}

\item{n}{(\code{number}):\cr number of patients.}

\item{p}{(\code{number}):\cr threshold that \code{P_E} is measured.}

\item{parE}{(\code{matrix}):\cr the beta parameters matrix, with \code{K} rows and 2 columns,
corresponding to the beta parameters of the \code{K} components.
Default is a uniform prior.}

\item{weights}{(\code{vector}):\cr The mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}

\item{betamixPost}{(\code{matrix}):\cr optional result of \verb{[getBetamixPost()]} in order
to speed up the computations. If supplied, this is directly used, bypassing
the other arguments (except \code{p} and \code{log.p} of course).}

\item{log.p}{(\code{number}):\cr whether to return the log of the probability}
}
\value{
The posterior probability that the response rate \code{P_E} is above \code{p}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute the posterior probability that the response probability \code{P_E}is above a threshold.
such that this posterior probability can be expressed as \code{Pr(P_E > p | data)}. Prior is
\code{P_E ~ sum(weights * beta(parE[, 1], parE[, 2]))}, i.e., a mixture of beta priors.
Default is one component only with uniform or \code{beta(1,1)}.

We observed \code{x} successes in n trials.

Posterior is again a mixture of beta priors, with updated mixture weights
and beta parameters.
}
\examples{
# Example taken from Lee and Liu (2008) :
# We observed 16 successes out of 23 patients
# We set a threshold of 0.60
# Assume a beta(0.6,0.4) prior for P_E
# Posterior will be a beta(16.6, 7.4), Pr(P_E > p | data) = 0.836

postprob(x = 16, n = 23, p = 0.60, par = c(0.6, 0.4))

# We could instead specify a mixture prior
# 2 component beta mixture prior, i.e., P_E ~ 0.6*beta(0.6,0.4) + 0.4*beta(1,1) and Pr(P_E > p | data) = 0.823
postprob(
  x = 16, n = 23, p = 0.60,
  par =
    rbind(
      c(0.6, 0.4),
      c(1, 1)
    ),
  weights = c(0.6, 0.4)
)
}
