% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprobBeta}
\alias{postprobBeta}
\title{Posterior Probability of Efficacy Given Beta Prior}
\usage{
postprobBeta(x, n, p, a = 1, b = 1)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of successes.}

\item{n}{(\code{number}):\cr number of patients.}

\item{p}{(\code{number}):\cr threshold set to compute posterior probability.}

\item{a}{(\code{matrix}):\cr first parameter \code{alpha} of the beta prior (successes).}

\item{b}{(\code{matrix}):\cr second parameter \code{beta} of the beta prior (failures).}
}
\value{
The posterior probability that the response rate P_E is above a threshold p.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute the posterior probability to be above threshold assuming a beta prior
on the response rate such that the response rate in the experimental or treatment \code{E} exceeds
the threshold set to compute the posterior probability \code{p} :  \code{Pr(P_E > p | data)}.
Prior is \code{P_E ~ beta(a, b)}, with default set to be a uniform or beta(1,1).

We observed \code{x} successes in n trials and so the posterior is
\code{P_E | data  ~ beta(a + x, b + n - x)}.
}
\examples{
# Example taken from Lee & Liu (2006)
# We observed 16 successes out of 23 patients # should we write this in the documentation
# We set a threshold of 0.60
# Assume a beta(0.6,0.4) prior for P_E
# Posterior will be a beta(16.6,22.8), Pr(P_E > p | data) = 0.8358808


# Example taken from Lee and Liu (2006)
postprobBeta(x = 16, n = 23, p = 0.60, a = 0.6, b = 0.4)
# Interpretation : The probability 16 of 23 successes is greater than 60 \% threshold is approximately 84 \%
}
