% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{postprobDist}
\alias{postprobDist}
\title{Compute the posterior probability with beta prior on SOC}
\usage{
postprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(1, 1),
  weights,
  parS = c(1, 1),
  weightsS
)
}
\arguments{
\item{x}{(\code{vector}):\cr vector of success counts in the treatment group. Vector of minimum length of 1.}

\item{n}{(\code{number}):\cr number of patients (in the treatment group).}

\item{xS}{(\code{vector}):\cr vector of success counts in the SOC group (default: 0). Vector of minimum length of 1.}

\item{nS}{(\code{number}):\cr number of patients in the SOC group (default: 0)}

\item{delta}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the SOC group (default: 0). Must be >= \code{0}.  see @note.
? response rate in delta proportion of the SOC non-responding patients.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the SOC non-responding patients. see @note.}

\item{parE}{(\code{matrix}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components. default is a
uniform prior.}

\item{weights}{(\code{matrix}):\cr the mixture weights of the beta mixture prior. Default are
uniform weights across mixture components.}

\item{parS}{(\code{matrix}):\cr beta parameters for the SOC group (default: uniform)}

\item{weightsS}{(\code{matrix}):\cr weights for the SOC group (default: uniform)}

\item{epsilon}{(\verb{# TODO}):\cr # do the integration. be careful to cover the region where there can}
}
\value{
The posterior probability
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Using the approach by Thall and Simon (Biometrics, 1994), evaluate the
posterior probability of having \code{Pr(P_E > P_S + delta | data)} (but see below
for relative delta margin). Both for the new treatment \code{E} as well as for the
Standard of Care (SOC): \code{S} data might be available. However the default is that no data is
available for the \code{SOC}, corresponding to the single arm trial situation. Note
that a uniform prior is the useful default for the treatment proportion,
while in the single arm trial an informative prior on the \code{SOC} proportion is
useful.

Beta mixture prior can be specified for the treatment \code{parE}
and \code{weights} parameters) and control proportion \code{parS} and
\code{weightsS} parameters), see \code{postprob} for details. Note
that being able to specify a beta mixture prior also on the control
treatment is e.g. important for the futility decision making (see the
\code{oc2} code).
}
\note{
on \verb{relative Delta} and \code{delta}.

When \code{relativeDelta = TRUE}, then it is assumed that \code{delta != 0}.

Given \code{delta != 0}, it is assumed that :
\itemize{
\item The posterior is calculate on \code{p + delta}.
\item \code{P_S} is the response rate of the SOC group or \code{p}.
\item \code{1-P_S} is the non-response rate of the SOC group.
\item The treatment group's response rate is assumed to be greater than
\code{P_S + (1-P_S)}.
\item Delta is assumed to be the difference between \code{P_S} and \code{1-P_S}
\item It is desirable to achieve a minimum response rate of
\code{P_S + (1 - P_S)*delta} response rate in the treatment group.
\item The expression for posterior probability is thus:
\code{Pr(P_E > P_S + (1 - P_S) * delta | data)}.
}

If \code{delta = 0} :
\itemize{
\item The default \code{delta = 0} is assumed.
\item \code{p} is the minimum response rate of which the posterior is calculate on.
}
}
\examples{
# example similar to Lee and Liu:
postprobDist(x = 16, n = 23, parE = c(0.6, 0.4), parS = c(0.6, 0.4))

## these two should give the same result:
postprobDist(
  x = 27, n = 34,
  xS = 0, nS = 0,
  delta = 0.15,
  parE = c(1, 1),
  parS = c(50007530, 49924090)
)

postprob(x = 27, n = 34, p = 0.65, parE = c(1, 1))
# ok, almost

# try out mixtures:
# play around with the beta parameters and weights to
# get a feeling.
# Note that very extreme beta parameters do no longer increase
# the return value, because then that mixture component is too
# unlikely a posteriori
postprobDist(
  x = 16, n = 23,
  parE =
    rbind(
      c(0.6, 0.4),
      c(10, 20)
    ),
  parS =
    rbind(
      c(0.6, 0.4),
      c(10, 10)
    ),
  weightsS = c(1, 3)
)
}
