% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprobDist.R
\name{postprobDist}
\alias{postprobDist}
\title{Compute the Posterior Probability with Beta Prior on \code{SOC}}
\usage{
postprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  delta,
  relativeDelta = FALSE,
  parE = c(1, 1),
  weights,
  parS = c(1, 1),
  weightsS
)
}
\arguments{
\item{x}{(\code{numeric}):\cr number of success counts in the treatment group.}

\item{n}{(\code{number}):\cr number of patients in the treatment group.}

\item{xS}{(\code{number}):\cr number of success counts in the SOC group.}

\item{nS}{(\code{number}):\cr number of patients in the SOC group.}

\item{delta}{(\code{number}):\cr margin by which the response rate in the treatment group should
be better than in the SOC group. Must be >= \code{0}. See note.}

\item{relativeDelta}{(\code{flag}):\cr If \code{TRUE}, then a \code{relativeDelta} is used. Represents that a minimum
response rate in magnitude of \code{delta} of the SOC non-responding patients. See note.}

\item{parE}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{E} group. See details.}

\item{weights}{(\code{numeric}):\cr the non-negative mixture weights of the beta mixture prior for group \code{E}. See details.}

\item{parS}{(\code{numeric} or \code{matrix}):\cr parameters for beta distribution. If it is a matrix, it needs to have 2 columns,
and each row corresponds to each component of a beta-mixture distribution
for the \code{S} group. See details.}

\item{weightsS}{(\code{numeric}):\cr the non-negative mixture weights of the beta mixture prior for group \code{S}. See details.}

\item{epsilon}{(\code{number}):\cr the smallest non-negative floating number to represent the lower bound for
the interval of integration.}
}
\value{
The posterior probability
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Using the approach by Thall and Simon (Biometrics, 1994), this evaluates the
posterior probability of achieving superior response rate in the treatment group compared to standard of care (SOC).
See note below for two formulations of the difference in response rates.
}
\details{
The beta mixture prior for the \code{E} arm requires argument \code{parE} and \code{weights}.
The beta mixture prior for the \code{S} arm requires argument \code{parS} and \code{weightsS}.
See \verb{[postprob()]} for details.

If a beta-mixture is used, by default, the weights are uniform across the components.
Weights can exceed 1, to which the algorithm will normalize the weights such that all weights sum to 1.
}
\note{
\subsection{Delta :}{

The desired improvement is denoted as \code{delta}. There are two options in using \code{delta}.
The absolute case when \code{relativeDelta = FALSE} and relative as when \code{relativeDelta = TRUE}.
\enumerate{
\item The absolute case is when we define an absolute delta, greater than \code{P_S},
the response rate of the \code{SOC} group such that
the posterior is \code{Pr(P_E > P_S + delta | data)}.
\item In the relative case, we suppose that the treatment group's
response rate is assumed to be greater than \code{P_S + (1-P_S) * delta} such that
the posterior is \code{Pr(P_E > P_S + (1 - P_S) * delta | data)}.
}
}
}
\examples{
# An example similar to Lee and Liu (2008).
postprobDist(
  x = 16,
  n = 23,
  parE = c(0.6, 0.4),
  parS = c(0.6, 0.4),
  delta = 0.1,
  relativeDelta = FALSE
)

# For a sequence of success outcomes for Experimental arm.
postprobDist(
  x = c(16, 17),
  n = 23,
  parE = c(0.6, 0.4),
  parS = c(0.6, 0.4),
  delta = 0.1,
  relativeDelta = FALSE
)

# When we use a relative difference and look at several possible number of responses.
postprobDist(
  x = 1:23,
  n = 23,
  parE = c(0.2, 0.8),
  parS = c(0.6, 0.4),
  delta = 0.1,
  relativeDelta = TRUE
)

# When we use beta mixtures for both the Experimental and SOC arms.
postprobDist(
  x = 16,
  n = 23,
  parE =
    rbind(
      c(0.6, 0.4),
      c(10, 20)
    ),
  parS =
    rbind(
      c(0.6, 0.4),
      c(10, 10)
    ),
  weightsS = c(1, 3),
  delta = 0.1
)

# Experimental arm only (strictly single arm trial), uniform prior in Experimental arm.
# Non-uniform Prior used for SOC arm as no precedent data.
postprobDist(
  x = 16,
  n = 23,
  xS = 0,
  nS = 0,
  delta = 0,
  relativeDelta = FALSE,
  parS = c(2, 3),
  weightsS = 1
)
# Experimental arm and SOC, uniform prior in both E and S arms, default setting used.
postprobDist(
  x = 16,
  n = 20,
  xS = 10,
  nS = 20,
  delta = 0,
  relativeDelta = FALSE,
  weightsS = 1
)

# Experimental and SOC arm, with beta mix prior for both arms.
# For each of the SOC arm is of 3 priors, therefore 3 sets of beta parameters, and 3 weights.
postprobDist(
  x = 16,
  n = 20,
  xS = 10,
  nS = 20,
  delta = 0.1,
  relativeDelta = TRUE,
  parE = rbind(c(1, 1), c(3, 4), c(8, 9)),
  weights = c(5, 3, 2),
  parS = rbind(c(4, 5), c(2, 3), c(4, 4)),
  weightsS = c(2, 5, 3)
)
}
