% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprobOld}
\alias{postprobOld}
\title{Compute the posterior probability to be above threshold assuming a beta prior
on the response rate}
\usage{
postprobOld(x, n, p, a = 1, b = 1)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of patients}

\item{p}{threshold}

\item{a}{first parameter of the beta prior (successes)}

\item{b}{second parameter of the beta prior (failures)}
}
\value{
The posterior probability that the response rate P_E is above a threshold p.
}
\description{
Computes the posterior probability Pr(P_E > p | data). Prior is P_E ~ beta(a, b),
with default set to be a uniform or beta(1,1).
}
\details{
We observed x successes in n trials and so the posterior is
P_E | data  ~ beta(a + x, b + n - x).
}
\examples{
## Example taken from Lee and Liu (????)
##
## We observed 16 successes out of 23 patients
## We set a threshold of 0.60
## Assume a beta(0.6,0.4) prior for P_E
## Posterior will be a beta(16.6,22.8), Pr(P_E > p | data) = 0.8358808
##
##

postprobOld(x = 16, n = 23, p = 0.60, a = 0.6, b = 0.4)
}
