% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprob.R
\name{predprob}
\alias{predprob}
\title{Predictive probability of trial success}
\usage{
predprob(x, n, Nmax, p, thetaT, parE = c(1, 1), weights)
}
\arguments{
\item{x}{(\code{number}):\cr number of successes.}

\item{n}{(\code{number}):\cr number of patients.}

\item{Nmax}{(\code{number}):\cr maximum number of patients at the end of the trial.}

\item{p}{(\code{number}):\cr threshold on the response rate.}

\item{thetaT}{(\code{number}):\cr threshold on the probability to be above p.}

\item{parE}{(\code{numeric}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}

\item{weights}{(\code{numeric}):\cr the mixture weights of the beta mixture prior.}
}
\value{
A \code{list} is returned with names \code{result} for predictive probability and
\code{table} of numeric values with counts, densities, posterior and result of indicator function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute the predictive probability that the trial will be
successful with fixed response rate threshold

Compute the predictive probability of trial success given current data.
Success means that at the end of the trial the posterior probability is
\code{Pr(P_E > p) >= thetaT},
where p is the fixed response rate threshold.
Then the predictive probability for success is:
\verb{pp = sum over i: Pr(Y=i|x,n)*I\{Pr(P_E > p|x,Y=i)>=thetaT\}}
where \code{Y} is the number of future responses in the treatment group and \code{x} is
the current number of responses in the treatment group out of n.
The prior for the response rate in the experimental arm is \code{P_E ~ beta(a, b)}.

A table with the following contents will be included in the return output :
\itemize{
\item \code{i}: \code{Y = i}, number of future successes in \code{Nmax-n} subjects.
\item \code{density}: \code{Pr(Y = i|x)} using beta-(mixture)-binomial distribution.
\item \code{posterior}: \code{Pr(P_E > p | x, Y=i)} using beta posterior.
\item \code{result_thetaT}: indicator \code{I(b>thetaT)}.
}
}
\examples{
# The original Lee and Liu (Table 1) example:
# Nmax = 40, x = 16, n = 23, beta(0.6,0.4) prior distribution,
# thetaT = 0.9. The control response rate is 60\%:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = c(0.6, 0.4)
)

# Lowering/Increasing the probability threshold thetaT of course increases
# /decreases the predictive probability of success, respectively:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.8,
  parE = c(0.6, 0.4)
)

predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.95,
  parE = c(0.6, 0.4)
)

# Mixed beta prior
predprob(
  x = 20, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = rbind(c(1, 1), c(25, 15)),
  weights = c(3, 1)
)
}
\references{
Lee, J. J., & Liu, D. D. (2008). A predictive probability
design for phase II cancer clinical trials. Clinical Trials, 5(2),
93-106. doi:10.1177/1740774508089279
}
