% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predprobDist.R
\name{predprobDist}
\alias{predprobDist}
\title{Compute the predictive probability that the trial will be
successful, with a prior distribution on the standard of care or \code{S}}
\usage{
predprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  Nmax,
  NmaxControl = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(a = 1, b = 1),
  weights,
  parS = c(a = 1, b = 1),
  weightsS,
  thetaT
)
}
\value{
A \code{list} is returned with names \code{result} for predictive probability including or separately a
\code{table} of numeric values with counts of responses in the remaining patients, \code{density} for
probabilities of these counts, \code{posterior} for corresponding probabilities to be above threshold,
and \code{success}for trial success indicators.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compute the predictive probability of trial success given current data.

Success means that at the end of the trial the posterior probability is:
\code{Pr(P_E > P_S + delta_0 | data) >= thetaT}.

Where \code{P_E} is the response rate of the treatment or \code{E}group and
\code{P_S} is the response rate of the standard of care of \code{S} group.

Then the predictive probability for success is:
\verb{pp = sum over i: Pr(Y = i | x , n)*I\{Pr(P_E > P_S + delta | x,Y=i) >= thetaT\}},
where \code{Y} is the number of future responses in the treatment group and \code{x} is
the current number of responses in the treatment group out of \code{n}.
Please see note in \code{\link[=postprobDist]{postprobDist()}} for definition of \code{delta} and \code{relativeDelta}.
Prior is \code{P_E ~ beta(a, b)} and uniform which is a \code{beta(1,1)}.
However, a beta mixture prior can also be specified. Analogously
for \code{P_S} either a classic beta prior or a beta mixture prior can be
specified.

Also data on the \code{S} might be available. Then the predictive probability is
more generally defined as :
\verb{pp = sum over i, j: Pr(Y = i | x, n)*Pr(Z = j | xS, nS )*I\{Pr(P_E > P_S + delta | x,xS, Y=i, Z=j ) >= thetaT\}}
where \code{Z} is the future number of responses in the \code{S} group, and \code{xS} is the
current number of responses in the \code{S} group.

In the case where \code{NmaxControl = 0}, a table with the following contents will be included in the return output :
\itemize{
\item \code{counts}: \code{Y = i}, number of future successes in \code{Nmax-n} subjects in ascending order.
\item \code{cumul_counts}: \code{Y = i}, number of future successes in \code{Nmax-n} subjects.
\item \code{density}: \code{Pr(Y = i|x)} using beta-(mixture)-binomial distribution.
\item \code{posterior}: \code{Pr(P_E > P_S + delta | x, Y = i)} using beta posterior.
\item \code{success}: indicator \code{I( b > thetaT )}.
}

If \code{NmaxControl} is not zero, i.e., when data on the control treatment
is available in this trial, then a list with will be included with the
following elements:
\itemize{
\item \code{counts}: \code{Y = i}, number of future successes in \code{Nmax-n} subjects in ascending order.
\item \code{cumul_counts}: \code{Y = i}, number of future successes in \code{Nmax-n} subjects.
\item \code{density}: \code{Pr(Y = i|x)} using beta-(mixture)-binomial distribution.
\item \code{posterior}: \code{Pr(P_E > P_S + delta | x, Y = i)} using beta posterior.
\item \code{success}: indicator \code{I( b > thetaT )}.
}
}
\examples{
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0, thetaT = 0.9,
  parE = c(0.6, 0.4), parS = c(7, 11)
)
# Specifying delta.
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0.1, thetaT = 0.9,
  parE = c(1, 1), parS = c(6, 11)
)

# Lowering theta increases predictive probability.
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0.1, thetaT = 0.5,
  parE = c(1, 1), parS = c(6, 11)
)

# Beta-mixture prior for experimental arm.
predprobDist(
  x = 16,
  n = 23,
  Nmax = 40,
  delta = 0.1,
  thetaT = 0.9,
  parE = rbind(c(1, 1), c(2, 6)),
  weights = c(2, 1),
  parS = c(6, 11)
)

# More extensive use of historical trials for control arm.
predprobDist(
  x = 16,
  n = 23,
  xS = 5,
  nS = 10,
  Nmax = 40,
  NmaxControl = 20,
  delta = 0.1,
  thetaT = 0.9,
  parE = rbind(c(1, 1), c(3, 5)),
  weights = c(2, 1),
  parS = rbind(c(1, 1), c(2, 6)),
  weightsS = c(2, 1)
)
}
\references{
Lee, J. J., & Liu, D. D. (2008). A predictive probability
design for phase II cancer clinical trials. Clinical Trials, 5(2),
93-106. doi:10.1177/1740774508089279
}
