% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiff.R
\name{qbetadiff}
\alias{qbetadiff}
\title{Beta difference Quantile Function}
\usage{
qbetadiff(p, parY, parX, eps = .Machine$double.eps)
}
\arguments{
\item{p}{(\code{number}):\cr vector of probabilities}

\item{parY}{(\code{numeric}):\cr two parameters of \code{Y}'s Beta distribution (Treatment)}

\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}

\item{eps}{(\code{number}):\cr lowest floating point number}
}
\value{
\code{qbetadiff}, the quantile function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the quantile of the Beta difference for a given probability.
}
\examples{
# qbetadiff ----

# We calculate quantile function of the Beta distribution of difference
# using the following parameters:
parX <- c(1, 52)
parY <- c(5.5, 20.5)

# Calculate quantile when at there is at least 20\% of difference.
test <- qbetadiff(
  p = 0.2,
  parY = parY,
  parX = parX
)
}
