% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{qbetaMix}
\alias{qbetaMix}
\title{Beta-mixture quantile function}
\usage{
qbetaMix(q, par, weights)
}
\arguments{
\item{q}{the required quantile}

\item{par}{the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components}

\item{weights}{the mixture weights of the beta mixture prior}
}
\value{
the abscissa
}
\description{
Note that \code{q} can be a vector.
}
