% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetabinom.R
\name{qbetaMix}
\alias{qbetaMix}
\title{Beta-mixture quantile function}
\usage{
qbetaMix(q, par, weights, lower.tail)
}
\arguments{
\item{q}{(\code{numeric}):\cr the required quantile.}

\item{par}{(\code{number}):\cr the beta parameters matrix, with K rows and 2 columns,
corresponding to the beta parameters of the K components.}

\item{weights}{(\code{matrix}):\cr the mixture weights of the beta mixture prior.}

\item{lower.tail}{(\code{flag}):\cr whether cdf at x taken at lower or upper tail}
}
\value{
the abscissa.
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Calculates the quantile where x support is at the intersection of cdf and quantile function at chosen quantile/s
}
\examples{
## Only 1 mixture component, i.e., weights = 1
qbetaMix(
  q = 0.60,
  par = rbind(c(0.2, 0.4)),
  weights = 1
)

## With 2 mixture components
qbetaMix(
  q = 0.6, par = rbind(c(0.2, 0.4), c(1, 1)),
  weights = c(0.6, 0.4)
)

## Can also specify q as a vector
qbetaMix(
  q = seq(0, 1, .01),
  par = rbind(c(0.2, 0.4), c(1, 1)),
  weights = c(0.6, 0.4)
)
}
