% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumBetaDiff.R
\name{sumBetaDiff}
\alias{sumBetaDiff}
\title{Mode and Credible Interval Calculation for The Difference between Two Beta Distributions.}
\usage{
sumBetaDiff(parX, parY, coverage = 0.9, go_cut, stop_cut)
}
\arguments{
\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}

\item{parY}{(\code{numeric}):\cr two parameters of \code{Y}'s Beta distribution (Treatment)}

\item{go_cut}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{stop_cut}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{conf_level}{(\code{numeric}):\cr coverage for credible interval}
}
\value{
\code{sumbetadiff} gives the mode, credible interval of the distribution function,
along with the probabilities of above \code{go_cut} and below \code{stop_cut}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to summarize the characters of a betadiff distribution \code{\link[=dbetadiff]{dbetadiff()}}.
May require use of random sample generator to calculate, use \code{\link[=set.seed]{set.seed()}} to reproduce results.
}
\examples{
parX <- c(1, 52) # Control group's parameters
parY <- c(5.5, 20.5) # Treatment group's parameters
sumBetadiff(
  parX = parX,
  parY = parY,
  coverage = 0.9,
  go_cut = 0.6,
  stop_cut = 0.2
)
}
