% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumbetadiff.R
\name{sumBetadiff}
\alias{sumBetadiff}
\title{Mode and Credible Interval Calculation for The Difference between Two Beta Distributions}
\usage{
sumBetadiff(parY, parX, level = 0.9, cutB, cutW)
}
\arguments{
\item{parY}{two parameters of Y's beta distribution (Treatment)}

\item{parX}{two parameters of X's beta distribution (Control)}

\item{level}{credible interval}

\item{cutB}{a cutoff value for the upper bound}

\item{cutW}{a cutoff value for the lower bound}
}
\value{
sumbetadiff gives the mode, credible interval of the distribution function,
along with the probabilities of above cutB and below cutW.
}
\description{
A function to summerize the characters of a betadiff distribution \code{\link{betadiff}}
}
\examples{
parX <- c(1, 52)
parY <- c(5.5, 20.5)
sumBetadiff(parX,parY,0.9,cutB=0.8,cutW=0.2)
}
