% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumbetadiff.R
\name{sumBetadiff}
\alias{sumBetadiff}
\title{Mode and Credible Interval Calculation for The Difference between Two Beta Distributions.}
\usage{
sumBetadiff(parY, parX, level = 0.9, cutB, cutW)
}
\arguments{
\item{parY}{(\code{numeric}):\cr two typedeters of Y's beta distribution (Treatment).}

\item{parX}{(\code{numeric}):\cr two parameters of X's beta distribution (Control).}

\item{level}{(``):\cr coverage for credible interval}

\item{cutB}{(``):\cr a cutoff value for the upper bound.}

\item{cutW}{(``):\cr a cutoff value for the lower bound.}
}
\value{
sumbetadiff gives the mode, credible interval of the distribution function,
along with the probabilities of above cutB and below cutW.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function to summarize the characters of a betadiff distribution \verb{[dbetadiff]}.
}
\examples{
parX <- c(1, 52)
parY <- c(5.5, 20.5)
sumBetadiff(parX, parY, 0.9, cutB = 0.8, cutW = 0.2)
}
