% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTable.R
\name{sumTable}
\alias{sumTable}
\title{Return summary statistics under one response scenario}
\usage{
sumTable(
  thisResp,
  TotalSample,
  go_cut,
  stop_cut,
  parX,
  YPri = c(0.5, 0.5),
  Round = 2
)
}
\arguments{
\item{thisResp}{(\code{number}):\cr number of responses}

\item{TotalSample}{(\code{number}):\cr sample size}

\item{go_cut}{(\code{number}):\cr a meaningful improvement threshold, the lower boundary of a meaningfully improvement in response rate}

\item{stop_cut}{(\code{number}):\cr a poor improvement threshold, the upper boundary of a meaningfully poor improvement in response rate}

\item{parX}{(\code{numeric}):\cr two parameters of \code{X}'s Beta distribution (Control)}

\item{YPri}{(\code{numeric}):\cr non-negative parameters of the beta prior of the treatment, default Beta(0.5,0.5)}

\item{Round}{(\code{number}):\cr Rounding of the output statistics}
}
\value{
A vector with the results.
}
\description{
This function will calculate the summary statistics for a specific response
outcome scenario.
}
\examples{
sumTable(2, 25, parX = c(1, 52), cut_B = 0.2, cut_W = 0.05, YPri = c(1, 1))

# for multiple response scenarios (e.g. 0 to 8 responses out of 25)
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    TotalSample = 25,
    parX = c(1, 52),
    go_cut = 0.2,
    stop_cut = 0.05
  )
)
summaries
}
