% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTable.R
\name{sumTable}
\alias{sumTable}
\title{Return summary statistics under one response scenario}
\usage{
sumTable(
  thisResp,
  TotalSample,
  go_cut,
  stop_cut,
  parX,
  YPri = c(0.5, 0.5),
  Round = 2
)
}
\arguments{
\item{thisResp}{number of responses}

\item{TotalSample}{sample size}

\item{parX}{non-negative parameters of the beta distribution of the control (posterior)}

\item{YPri}{non-negative parameters of the beta prior of the treatment, default Beta(0.5,0.5)}

\item{Round}{Rounding of the output statistics}

\item{cut_B}{a meaningful improvement threshold: at least cut_B (e.g. 15\%) improvement}

\item{cut_W}{a poor improvement threshold: at most cut_W (e.g. 5\%) improvement}
}
\value{
A vector with the results.
}
\description{
This function will calculate the summary statistics for a specific response
outcome scenario.
}
\examples{
sumTable(2, 25, parX = c(1, 52), cut_B = 0.2, cut_W = 0.05, YPri = c(1, 1))

# for multiple response scenarios (e.g. 0 to 8 responses out of 25)
summaries <- do.call(
  cbind,
  lapply(c(0:8),
    sumTable,
    TotalSample = 25,
    parX = c(1, 52),
    cut_B = 0.2,
    cut_W = 0.05
  )
)
summaries
}
